@@ -9,99 +9,71 @@
     "homepage": "http://framework.zend.com/",
     "license": "BSD-3-Clause",
     "require": {
-        "php": ">=5.3.23",
+        "php": "^5.5 || ^7.0",
+        "zendframework/zend-authentication": "^2.5",
+        "zendframework/zend-barcode": "^2.5",
+        "zendframework/zend-cache": "^2.5",
+        "zendframework/zend-captcha": "^2.5",
+        "zendframework/zend-code": "^2.5",
+        "zendframework/zend-config": "^2.5",
+        "zendframework/zend-console": "^2.5",
+        "zendframework/zend-crypt": "^2.5",
+        "zendframework/zend-db": "^2.5",
+        "zendframework/zend-debug": "^2.5",
+        "zendframework/zend-di": "^2.5",
+        "zendframework/zend-dom": "^2.5",
+        "zendframework/zend-escaper": "^2.5",
+        "zendframework/zend-eventmanager": "^2.5",
+        "zendframework/zend-feed": "^2.5",
+        "zendframework/zend-file": "^2.5",
+        "zendframework/zend-filter": "^2.5",
+        "zendframework/zend-form": "^2.5",
+        "zendframework/zend-http": "^2.5",
+        "zendframework/zend-i18n": "^2.5",
+        "zendframework/zend-i18n-resources": "^2.5",
+        "zendframework/zend-inputfilter": "^2.5",
+        "zendframework/zend-json": "^2.5",
+        "zendframework/zend-loader": "^2.5",
+        "zendframework/zend-log": "^2.5",
+        "zendframework/zend-mail": "^2.5",
+        "zendframework/zend-math": "^2.5",
+        "zendframework/zend-memory": "^2.5",
+        "zendframework/zend-mime": "^2.5",
+        "zendframework/zend-modulemanager": "^2.5",
+        "zendframework/zend-mvc": "^2.5",
+        "zendframework/zend-navigation": "^2.5",
+        "zendframework/zend-paginator": "^2.5",
+        "zendframework/zend-permissions-acl": "^2.5",
+        "zendframework/zend-permissions-rbac": "^2.5",
+        "zendframework/zend-progressbar": "^2.5",
+        "zendframework/zend-serializer": "^2.5",
+        "zendframework/zend-server": "^2.5",
+        "zendframework/zend-servicemanager": "^2.5",
+        "zendframework/zend-session": "^2.5",
+        "zendframework/zend-soap": "^2.5",
+        "zendframework/zend-stdlib": "^2.5",
+        "zendframework/zend-tag": "^2.5",
+        "zendframework/zend-test": "^2.5",
+        "zendframework/zend-text": "^2.5",
+        "zendframework/zend-uri": "^2.5",
+        "zendframework/zend-validator": "^2.5",
+        "zendframework/zend-version": "^2.5",
+        "zendframework/zend-view": "^2.5",
+        "zendframework/zend-xmlrpc": "^2.5",
         "zendframework/zendxml": "^1.0.1"
     },
-    "require-dev": {
-        "doctrine/annotations": "~1.0",
-        "ircmaxell/random-lib": "~1.1",
-        "mikey179/vfsStream": "~1.2",
-        "fabpot/php-cs-fixer": "~1.0",
-        "phpunit/PHPUnit": "~4.6",
-        "satooshi/php-coveralls": "dev-master",
-        "phpunit/phpcov": "~2.0"
-    },
     "suggest": {
-        "ext-intl": "ext/intl for i18n features (included in default builds of PHP)",
-        "doctrine/annotations": "Doctrine Annotations >=1.0 for annotation features",
-        "ircmaxell/random-lib": "Fallback random byte generator for Zend\\Math\\Rand if OpenSSL/Mcrypt extensions are unavailable",
-        "ocramius/proxy-manager": "ProxyManager 0.5.* to handle lazy initialization of services",
-        "zendframework/zendpdf": "ZendPdf for creating PDF representations of barcodes",
-        "zendframework/zendservice-recaptcha": "ZendService\\ReCaptcha for rendering ReCaptchas in Zend\\Captcha and/or Zend\\Form"
-    },
-    "autoload": {
-        "psr-4": {
-            "Zend\\": "library/Zend/"
-        }
-    },
-    "autoload-dev": {
-        "psr-0": "using PSR-0 for tests, as we have test assets with underscores",
-        "psr-0": {
-            "ZendTest\\": "tests/"
-        }
+        "zendframework/zend-ldap": "zend-ldap component ^2.5, if you need LDAP features"
     },
     "bin": [
         "bin/classmap_generator.php",
         "bin/pluginmap_generator.php",
         "bin/templatemap_generator.php"
     ],
-    "replace": {
-        "zendframework/zend-authentication": "self.version",
-        "zendframework/zend-barcode": "self.version",
-        "zendframework/zend-cache": "self.version",
-        "zendframework/zend-captcha": "self.version",
-        "zendframework/zend-code": "self.version",
-        "zendframework/zend-config": "self.version",
-        "zendframework/zend-console": "self.version",
-        "zendframework/zend-crypt": "self.version",
-        "zendframework/zend-db": "self.version",
-        "zendframework/zend-debug": "self.version",
-        "zendframework/zend-di": "self.version",
-        "zendframework/zend-dom": "self.version",
-        "zendframework/zend-escaper": "self.version",
-        "zendframework/zend-eventmanager": "self.version",
-        "zendframework/zend-feed": "self.version",
-        "zendframework/zend-file": "self.version",
-        "zendframework/zend-filter": "self.version",
-        "zendframework/zend-form": "self.version",
-        "zendframework/zend-http": "self.version",
-        "zendframework/zend-i18n": "self.version",
-        "zendframework/zend-inputfilter": "self.version",
-        "zendframework/zend-json": "self.version",
-        "zendframework/zend-ldap": "self.version",
-        "zendframework/zend-loader": "self.version",
-        "zendframework/zend-log": "self.version",
-        "zendframework/zend-mail": "self.version",
-        "zendframework/zend-math": "self.version",
-        "zendframework/zend-memory": "self.version",
-        "zendframework/zend-mime": "self.version",
-        "zendframework/zend-modulemanager": "self.version",
-        "zendframework/zend-mvc": "self.version",
-        "zendframework/zend-navigation": "self.version",
-        "zendframework/zend-paginator": "self.version",
-        "zendframework/zend-permissions-acl": "self.version",
-        "zendframework/zend-permissions-rbac": "self.version",
-        "zendframework/zend-progressbar": "self.version",
-        "zendframework/zend-resources": "self.version",
-        "zendframework/zend-serializer": "self.version",
-        "zendframework/zend-server": "self.version",
-        "zendframework/zend-servicemanager": "self.version",
-        "zendframework/zend-session": "self.version",
-        "zendframework/zend-soap": "self.version",
-        "zendframework/zend-stdlib": "self.version",
-        "zendframework/zend-tag": "self.version",
-        "zendframework/zend-test": "self.version",
-        "zendframework/zend-text": "self.version",
-        "zendframework/zend-uri": "self.version",
-        "zendframework/zend-validator": "self.version",
-        "zendframework/zend-version": "self.version",
-        "zendframework/zend-view": "self.version",
-        "zendframework/zend-xmlrpc": "self.version"
-    },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.4-dev",
-            "dev-develop": "2.5-dev"
+            "dev-master": "2.5-dev",
+            "dev-develop": "2.6-dev"
         }
     }
 }
