@@ -22,20 +22,20 @@ class DumperTest extends TestCase
     protected $dumper;
     protected $path;
 
-    protected $array = array(
+    protected $array = [
         '' => 'bar',
         'foo' => '#bar',
-        'foo\'bar' => array(),
-        'bar' => array(1, 'foo'),
-        'foobar' => array(
+        'foo\'bar' => [],
+        'bar' => [1, 'foo'],
+        'foobar' => [
             'foo' => 'bar',
-            'bar' => array(1, 'foo'),
-            'foobar' => array(
+            'bar' => [1, 'foo'],
+            'foobar' => [
                 'foo' => 'bar',
-                'bar' => array(1, 'foo'),
-            ),
-        ),
-    );
+                'bar' => [1, 'foo'],
+            ],
+        ],
+    ];
 
     protected function setUp()
     {
@@ -179,14 +179,14 @@ EOF;
 
     public function testObjectSupportEnabled()
     {
-        $dump = $this->dumper->dump(array('foo' => new A(), 'bar' => 1), 0, 0, Yaml::DUMP_OBJECT);
+        $dump = $this->dumper->dump(['foo' => new A(), 'bar' => 1], 0, 0, Yaml::DUMP_OBJECT);
 
         $this->assertEquals('{ foo: !php/object \'O:30:"Symfony\Component\Yaml\Tests\A":1:{s:1:"a";s:3:"foo";}\', bar: 1 }', $dump, '->dump() is able to dump objects');
     }
 
     public function testObjectSupportDisabledButNoExceptions()
     {
-        $dump = $this->dumper->dump(array('foo' => new A(), 'bar' => 1));
+        $dump = $this->dumper->dump(['foo' => new A(), 'bar' => 1]);
 
         $this->assertEquals('{ foo: null, bar: 1 }', $dump, '->dump() does not dump objects when disabled');
     }
@@ -196,7 +196,7 @@ EOF;
      */
     public function testObjectSupportDisabledWithExceptions()
     {
-        $this->dumper->dump(array('foo' => new A(), 'bar' => 1), 0, 0, Yaml::DUMP_EXCEPTION_ON_INVALID_TYPE);
+        $this->dumper->dump(['foo' => new A(), 'bar' => 1], 0, 0, Yaml::DUMP_EXCEPTION_ON_INVALID_TYPE);
     }
 
     /**
@@ -209,27 +209,27 @@ EOF;
 
     public function getEscapeSequences()
     {
-        return array(
-            'empty string' => array('', "''"),
-            'null' => array("\x0", '"\\0"'),
-            'bell' => array("\x7", '"\\a"'),
-            'backspace' => array("\x8", '"\\b"'),
-            'horizontal-tab' => array("\t", '"\\t"'),
-            'line-feed' => array("\n", '"\\n"'),
-            'vertical-tab' => array("\v", '"\\v"'),
-            'form-feed' => array("\xC", '"\\f"'),
-            'carriage-return' => array("\r", '"\\r"'),
-            'escape' => array("\x1B", '"\\e"'),
-            'space' => array(' ', "' '"),
-            'double-quote' => array('"', "'\"'"),
-            'slash' => array('/', '/'),
-            'backslash' => array('\\', '\\'),
-            'next-line' => array("\xC2\x85", '"\\N"'),
-            'non-breaking-space' => array("\xc2\xa0", '"\\_"'),
-            'line-separator' => array("\xE2\x80\xA8", '"\\L"'),
-            'paragraph-separator' => array("\xE2\x80\xA9", '"\\P"'),
-            'colon' => array(':', "':'"),
-        );
+        return [
+            'empty string' => ['', "''"],
+            'null' => ["\x0", '"\\0"'],
+            'bell' => ["\x7", '"\\a"'],
+            'backspace' => ["\x8", '"\\b"'],
+            'horizontal-tab' => ["\t", '"\\t"'],
+            'line-feed' => ["\n", '"\\n"'],
+            'vertical-tab' => ["\v", '"\\v"'],
+            'form-feed' => ["\xC", '"\\f"'],
+            'carriage-return' => ["\r", '"\\r"'],
+            'escape' => ["\x1B", '"\\e"'],
+            'space' => [' ', "' '"],
+            'double-quote' => ['"', "'\"'"],
+            'slash' => ['/', '/'],
+            'backslash' => ['\\', '\\'],
+            'next-line' => ["\xC2\x85", '"\\N"'],
+            'non-breaking-space' => ["\xc2\xa0", '"\\_"'],
+            'line-separator' => ["\xE2\x80\xA8", '"\\L"'],
+            'paragraph-separator' => ["\xE2\x80\xA9", '"\\P"'],
+            'colon' => [':', "':'"],
+        ];
     }
 
     public function testBinaryDataIsDumpedBase64Encoded()
@@ -237,7 +237,7 @@ EOF;
         $binaryData = file_get_contents(__DIR__.'/Fixtures/arrow.gif');
         $expected = '{ data: !!binary '.base64_encode($binaryData).' }';
 
-        $this->assertSame($expected, $this->dumper->dump(array('data' => $binaryData)));
+        $this->assertSame($expected, $this->dumper->dump(['data' => $binaryData]));
     }
 
     public function testNonUtf8DataIsDumpedBase64Encoded()
@@ -258,18 +258,18 @@ EOF;
 
     public function objectAsMapProvider()
     {
-        $tests = array();
+        $tests = [];
 
         $bar = new \stdClass();
         $bar->class = 'classBar';
-        $bar->args = array('bar');
+        $bar->args = ['bar'];
         $zar = new \stdClass();
         $foo = new \stdClass();
         $foo->bar = $bar;
         $foo->zar = $zar;
         $object = new \stdClass();
         $object->foo = $foo;
-        $tests['stdClass'] = array($object, $object);
+        $tests['stdClass'] = [$object, $object];
 
         $arrayObject = new \ArrayObject();
         $arrayObject['foo'] = 'bar';
@@ -277,19 +277,19 @@ EOF;
         $parsedArrayObject = new \stdClass();
         $parsedArrayObject->foo = 'bar';
         $parsedArrayObject->baz = 'foobar';
-        $tests['ArrayObject'] = array($arrayObject, $parsedArrayObject);
+        $tests['ArrayObject'] = [$arrayObject, $parsedArrayObject];
 
         $a = new A();
-        $tests['arbitrary-object'] = array($a, null);
+        $tests['arbitrary-object'] = [$a, null];
 
         return $tests;
     }
 
     public function testDumpingArrayObjectInstancesRespectsInlineLevel()
     {
-        $deep = new \ArrayObject(array('deep1' => 'd', 'deep2' => 'e'));
-        $inner = new \ArrayObject(array('inner1' => 'b', 'inner2' => 'c', 'inner3' => $deep));
-        $outer = new \ArrayObject(array('outer1' => 'a', 'outer2' => $inner));
+        $deep = new \ArrayObject(['deep1' => 'd', 'deep2' => 'e']);
+        $inner = new \ArrayObject(['inner1' => 'b', 'inner2' => 'c', 'inner3' => $deep]);
+        $outer = new \ArrayObject(['outer1' => 'a', 'outer2' => $inner]);
 
         $yaml = $this->dumper->dump($outer, 2, 0, Yaml::DUMP_OBJECT_AS_MAP);
 
@@ -306,9 +306,9 @@ YAML;
 
     public function testDumpingArrayObjectInstancesWithNumericKeysInlined()
     {
-        $deep = new \ArrayObject(array('d', 'e'));
-        $inner = new \ArrayObject(array('b', 'c', $deep));
-        $outer = new \ArrayObject(array('a', $inner));
+        $deep = new \ArrayObject(['d', 'e']);
+        $inner = new \ArrayObject(['b', 'c', $deep]);
+        $outer = new \ArrayObject(['a', $inner]);
 
         $yaml = $this->dumper->dump($outer, 0, 0, Yaml::DUMP_OBJECT_AS_MAP);
         $expected = <<<YAML
@@ -319,9 +319,9 @@ YAML;
 
     public function testDumpingArrayObjectInstancesWithNumericKeysRespectsInlineLevel()
     {
-        $deep = new \ArrayObject(array('d', 'e'));
-        $inner = new \ArrayObject(array('b', 'c', $deep));
-        $outer = new \ArrayObject(array('a', $inner));
+        $deep = new \ArrayObject(['d', 'e']);
+        $inner = new \ArrayObject(['b', 'c', $deep]);
+        $outer = new \ArrayObject(['a', $inner]);
         $yaml = $this->dumper->dump($outer, 2, 0, Yaml::DUMP_OBJECT_AS_MAP);
         $expected = <<<YAML
 0: a
@@ -374,34 +374,34 @@ YAML;
 
     public function testDumpMultiLineStringAsScalarBlock()
     {
-        $data = array(
-            'data' => array(
+        $data = [
+            'data' => [
                 'single_line' => 'foo bar baz',
                 'multi_line' => "foo\nline with trailing spaces:\n  \nbar\ninteger like line:\n123456789\nempty line:\n\nbaz",
                 'multi_line_with_carriage_return' => "foo\nbar\r\nbaz",
-                'nested_inlined_multi_line_string' => array(
+                'nested_inlined_multi_line_string' => [
                     'inlined_multi_line' => "foo\nbar\r\nempty line:\n\nbaz",
-                ),
-            ),
-        );
+                ],
+            ],
+        ];
 
         $this->assertSame(file_get_contents(__DIR__.'/Fixtures/multiple_lines_as_literal_block.yml'), $this->dumper->dump($data, 2, 0, Yaml::DUMP_MULTI_LINE_LITERAL_BLOCK));
     }
 
     public function testDumpMultiLineStringAsScalarBlockWhenFirstLineHasLeadingSpace()
     {
-        $data = array(
-            'data' => array(
+        $data = [
+            'data' => [
                 'multi_line' => "    the first line has leading spaces\nThe second line does not.",
-            ),
-        );
+            ],
+        ];
 
         $this->assertSame(file_get_contents(__DIR__.'/Fixtures/multiple_lines_as_literal_block_leading_space_in_first_line.yml'), $this->dumper->dump($data, 2, 0, Yaml::DUMP_MULTI_LINE_LITERAL_BLOCK));
     }
 
     public function testCarriageReturnIsMaintainedWhenDumpingAsMultiLineLiteralBlock()
     {
-        $this->assertSame("- \"a\\r\\nb\\nc\"\n", $this->dumper->dump(array("a\r\nb\nc"), 2, 0, Yaml::DUMP_MULTI_LINE_LITERAL_BLOCK));
+        $this->assertSame("- \"a\\r\\nb\\nc\"\n", $this->dumper->dump(["a\r\nb\nc"], 2, 0, Yaml::DUMP_MULTI_LINE_LITERAL_BLOCK));
     }
 
     /**
