@@ -54,13 +54,13 @@ class InlineTest extends TestCase
 
     public function getTestsForParsePhpConstants()
     {
-        return array(
-            array('!php/const Symfony\Component\Yaml\Yaml::PARSE_CONSTANT', Yaml::PARSE_CONSTANT),
-            array('!php/const PHP_INT_MAX', PHP_INT_MAX),
-            array('[!php/const PHP_INT_MAX]', array(PHP_INT_MAX)),
-            array('{ foo: !php/const PHP_INT_MAX }', array('foo' => PHP_INT_MAX)),
-            array('!php/const NULL', null),
-        );
+        return [
+            ['!php/const Symfony\Component\Yaml\Yaml::PARSE_CONSTANT', Yaml::PARSE_CONSTANT],
+            ['!php/const PHP_INT_MAX', PHP_INT_MAX],
+            ['[!php/const PHP_INT_MAX]', [PHP_INT_MAX]],
+            ['{ foo: !php/const PHP_INT_MAX }', ['foo' => PHP_INT_MAX]],
+            ['!php/const NULL', null],
+        ];
     }
 
     /**
@@ -99,7 +99,7 @@ class InlineTest extends TestCase
         }
 
         try {
-            $requiredLocales = array('fr_FR.UTF-8', 'fr_FR.UTF8', 'fr_FR.utf-8', 'fr_FR.utf8', 'French_France.1252');
+            $requiredLocales = ['fr_FR.UTF-8', 'fr_FR.UTF8', 'fr_FR.utf-8', 'fr_FR.utf8', 'French_France.1252'];
             if (false === setlocale(LC_NUMERIC, $requiredLocales)) {
                 $this->markTestSkipped('Could not set any of required locales: '.implode(', ', $requiredLocales));
             }
@@ -200,31 +200,31 @@ class InlineTest extends TestCase
      */
     public function testParseReferences($yaml, $expected)
     {
-        $this->assertSame($expected, Inline::parse($yaml, 0, array('var' => 'var-value')));
+        $this->assertSame($expected, Inline::parse($yaml, 0, ['var' => 'var-value']));
     }
 
     public function getDataForParseReferences()
     {
-        return array(
-            'scalar' => array('*var', 'var-value'),
-            'list' => array('[ *var ]', array('var-value')),
-            'list-in-list' => array('[[ *var ]]', array(array('var-value'))),
-            'map-in-list' => array('[ { key: *var } ]', array(array('key' => 'var-value'))),
-            'embedded-mapping-in-list' => array('[ key: *var ]', array(array('key' => 'var-value'))),
-            'map' => array('{ key: *var }', array('key' => 'var-value')),
-            'list-in-map' => array('{ key: [*var] }', array('key' => array('var-value'))),
-            'map-in-map' => array('{ foo: { bar: *var } }', array('foo' => array('bar' => 'var-value'))),
-        );
+        return [
+            'scalar' => ['*var', 'var-value'],
+            'list' => ['[ *var ]', ['var-value']],
+            'list-in-list' => ['[[ *var ]]', [['var-value']]],
+            'map-in-list' => ['[ { key: *var } ]', [['key' => 'var-value']]],
+            'embedded-mapping-in-list' => ['[ key: *var ]', [['key' => 'var-value']]],
+            'map' => ['{ key: *var }', ['key' => 'var-value']],
+            'list-in-map' => ['{ key: [*var] }', ['key' => ['var-value']]],
+            'map-in-map' => ['{ foo: { bar: *var } }', ['foo' => ['bar' => 'var-value']]],
+        ];
     }
 
     public function testParseMapReferenceInSequence()
     {
-        $foo = array(
+        $foo = [
             'a' => 'Steve',
             'b' => 'Clark',
             'c' => 'Brian',
-        );
-        $this->assertSame(array($foo), Inline::parse('[*foo]', 0, array('foo' => $foo)));
+        ];
+        $this->assertSame([$foo], Inline::parse('[*foo]', 0, ['foo' => $foo]));
     }
 
     /**
@@ -262,7 +262,7 @@ class InlineTest extends TestCase
 
     public function getReservedIndicators()
     {
-        return array(array('@'), array('`'));
+        return [['@'], ['`']];
     }
 
     /**
@@ -282,7 +282,7 @@ class InlineTest extends TestCase
 
     public function getScalarIndicators()
     {
-        return array(array('|'), array('>'), array('%'));
+        return [['|'], ['>'], ['%']];
     }
 
     /**
@@ -295,240 +295,240 @@ class InlineTest extends TestCase
 
     public function getDataForIsHash()
     {
-        return array(
-            array(array(), false),
-            array(array(1, 2, 3), false),
-            array(array(2 => 1, 1 => 2, 0 => 3), true),
-            array(array('foo' => 1, 'bar' => 2), true),
-        );
+        return [
+            [[], false],
+            [[1, 2, 3], false],
+            [[2 => 1, 1 => 2, 0 => 3], true],
+            [['foo' => 1, 'bar' => 2], true],
+        ];
     }
 
     public function getTestsForParse()
     {
-        return array(
-            array('', ''),
-            array('null', null),
-            array('false', false),
-            array('true', true),
-            array('12', 12),
-            array('-12', -12),
-            array('1_2', 12),
-            array('_12', '_12'),
-            array('12_', 12),
-            array('"quoted string"', 'quoted string'),
-            array("'quoted string'", 'quoted string'),
-            array('12.30e+02', 12.30e+02),
-            array('123.45_67', 123.4567),
-            array('0x4D2', 0x4D2),
-            array('0x_4_D_2_', 0x4D2),
-            array('02333', 02333),
-            array('0_2_3_3_3', 02333),
-            array('.Inf', -log(0)),
-            array('-.Inf', log(0)),
-            array("'686e444'", '686e444'),
-            array('686e444', 646e444),
-            array('123456789123456789123456789123456789', '123456789123456789123456789123456789'),
-            array('"foo\r\nbar"', "foo\r\nbar"),
-            array("'foo#bar'", 'foo#bar'),
-            array("'foo # bar'", 'foo # bar'),
-            array("'#cfcfcf'", '#cfcfcf'),
-            array('::form_base.html.twig', '::form_base.html.twig'),
+        return [
+            ['', ''],
+            ['null', null],
+            ['false', false],
+            ['true', true],
+            ['12', 12],
+            ['-12', -12],
+            ['1_2', 12],
+            ['_12', '_12'],
+            ['12_', 12],
+            ['"quoted string"', 'quoted string'],
+            ["'quoted string'", 'quoted string'],
+            ['12.30e+02', 12.30e+02],
+            ['123.45_67', 123.4567],
+            ['0x4D2', 0x4D2],
+            ['0x_4_D_2_', 0x4D2],
+            ['02333', 02333],
+            ['0_2_3_3_3', 02333],
+            ['.Inf', -log(0)],
+            ['-.Inf', log(0)],
+            ["'686e444'", '686e444'],
+            ['686e444', 646e444],
+            ['123456789123456789123456789123456789', '123456789123456789123456789123456789'],
+            ['"foo\r\nbar"', "foo\r\nbar"],
+            ["'foo#bar'", 'foo#bar'],
+            ["'foo # bar'", 'foo # bar'],
+            ["'#cfcfcf'", '#cfcfcf'],
+            ['::form_base.html.twig', '::form_base.html.twig'],
 
             // Pre-YAML-1.2 booleans
-            array("'y'", 'y'),
-            array("'n'", 'n'),
-            array("'yes'", 'yes'),
-            array("'no'", 'no'),
-            array("'on'", 'on'),
-            array("'off'", 'off'),
+            ["'y'", 'y'],
+            ["'n'", 'n'],
+            ["'yes'", 'yes'],
+            ["'no'", 'no'],
+            ["'on'", 'on'],
+            ["'off'", 'off'],
 
-            array('2007-10-30', gmmktime(0, 0, 0, 10, 30, 2007)),
-            array('2007-10-30T02:59:43Z', gmmktime(2, 59, 43, 10, 30, 2007)),
-            array('2007-10-30 02:59:43 Z', gmmktime(2, 59, 43, 10, 30, 2007)),
-            array('1960-10-30 02:59:43 Z', gmmktime(2, 59, 43, 10, 30, 1960)),
-            array('1730-10-30T02:59:43Z', gmmktime(2, 59, 43, 10, 30, 1730)),
+            ['2007-10-30', gmmktime(0, 0, 0, 10, 30, 2007)],
+            ['2007-10-30T02:59:43Z', gmmktime(2, 59, 43, 10, 30, 2007)],
+            ['2007-10-30 02:59:43 Z', gmmktime(2, 59, 43, 10, 30, 2007)],
+            ['1960-10-30 02:59:43 Z', gmmktime(2, 59, 43, 10, 30, 1960)],
+            ['1730-10-30T02:59:43Z', gmmktime(2, 59, 43, 10, 30, 1730)],
 
-            array('"a \\"string\\" with \'quoted strings inside\'"', 'a "string" with \'quoted strings inside\''),
-            array("'a \"string\" with ''quoted strings inside'''", 'a "string" with \'quoted strings inside\''),
+            ['"a \\"string\\" with \'quoted strings inside\'"', 'a "string" with \'quoted strings inside\''],
+            ["'a \"string\" with ''quoted strings inside'''", 'a "string" with \'quoted strings inside\''],
 
             // sequences
             // urls are no key value mapping. see #3609. Valid yaml "key: value" mappings require a space after the colon
-            array('[foo, http://urls.are/no/mappings, false, null, 12]', array('foo', 'http://urls.are/no/mappings', false, null, 12)),
-            array('[  foo  ,   bar , false  ,  null     ,  12  ]', array('foo', 'bar', false, null, 12)),
-            array('[\'foo,bar\', \'foo bar\']', array('foo,bar', 'foo bar')),
+            ['[foo, http://urls.are/no/mappings, false, null, 12]', ['foo', 'http://urls.are/no/mappings', false, null, 12]],
+            ['[  foo  ,   bar , false  ,  null     ,  12  ]', ['foo', 'bar', false, null, 12]],
+            ['[\'foo,bar\', \'foo bar\']', ['foo,bar', 'foo bar']],
 
             // mappings
-            array('{foo: bar,bar: foo,"false": false, "null": null,integer: 12}', array('foo' => 'bar', 'bar' => 'foo', 'false' => false, 'null' => null, 'integer' => 12)),
-            array('{ foo  : bar, bar : foo, "false"  :   false,  "null"  :   null,  integer :  12  }', array('foo' => 'bar', 'bar' => 'foo', 'false' => false, 'null' => null, 'integer' => 12)),
-            array('{foo: \'bar\', bar: \'foo: bar\'}', array('foo' => 'bar', 'bar' => 'foo: bar')),
-            array('{\'foo\': \'bar\', "bar": \'foo: bar\'}', array('foo' => 'bar', 'bar' => 'foo: bar')),
-            array('{\'foo\'\'\': \'bar\', "bar\"": \'foo: bar\'}', array('foo\'' => 'bar', 'bar"' => 'foo: bar')),
-            array('{\'foo: \': \'bar\', "bar: ": \'foo: bar\'}', array('foo: ' => 'bar', 'bar: ' => 'foo: bar')),
-            array('{"foo:bar": "baz"}', array('foo:bar' => 'baz')),
-            array('{"foo":"bar"}', array('foo' => 'bar')),
+            ['{foo: bar,bar: foo,"false": false, "null": null,integer: 12}', ['foo' => 'bar', 'bar' => 'foo', 'false' => false, 'null' => null, 'integer' => 12]],
+            ['{ foo  : bar, bar : foo, "false"  :   false,  "null"  :   null,  integer :  12  }', ['foo' => 'bar', 'bar' => 'foo', 'false' => false, 'null' => null, 'integer' => 12]],
+            ['{foo: \'bar\', bar: \'foo: bar\'}', ['foo' => 'bar', 'bar' => 'foo: bar']],
+            ['{\'foo\': \'bar\', "bar": \'foo: bar\'}', ['foo' => 'bar', 'bar' => 'foo: bar']],
+            ['{\'foo\'\'\': \'bar\', "bar\"": \'foo: bar\'}', ['foo\'' => 'bar', 'bar"' => 'foo: bar']],
+            ['{\'foo: \': \'bar\', "bar: ": \'foo: bar\'}', ['foo: ' => 'bar', 'bar: ' => 'foo: bar']],
+            ['{"foo:bar": "baz"}', ['foo:bar' => 'baz']],
+            ['{"foo":"bar"}', ['foo' => 'bar']],
 
             // nested sequences and mappings
-            array('[foo, [bar, foo]]', array('foo', array('bar', 'foo'))),
-            array('[foo, {bar: foo}]', array('foo', array('bar' => 'foo'))),
-            array('{ foo: {bar: foo} }', array('foo' => array('bar' => 'foo'))),
-            array('{ foo: [bar, foo] }', array('foo' => array('bar', 'foo'))),
-            array('{ foo:{bar: foo} }', array('foo' => array('bar' => 'foo'))),
-            array('{ foo:[bar, foo] }', array('foo' => array('bar', 'foo'))),
+            ['[foo, [bar, foo]]', ['foo', ['bar', 'foo']]],
+            ['[foo, {bar: foo}]', ['foo', ['bar' => 'foo']]],
+            ['{ foo: {bar: foo} }', ['foo' => ['bar' => 'foo']]],
+            ['{ foo: [bar, foo] }', ['foo' => ['bar', 'foo']]],
+            ['{ foo:{bar: foo} }', ['foo' => ['bar' => 'foo']]],
+            ['{ foo:[bar, foo] }', ['foo' => ['bar', 'foo']]],
 
-            array('[  foo, [  bar, foo  ]  ]', array('foo', array('bar', 'foo'))),
+            ['[  foo, [  bar, foo  ]  ]', ['foo', ['bar', 'foo']]],
 
-            array('[{ foo: {bar: foo} }]', array(array('foo' => array('bar' => 'foo')))),
+            ['[{ foo: {bar: foo} }]', [['foo' => ['bar' => 'foo']]]],
 
-            array('[foo, [bar, [foo, [bar, foo]], foo]]', array('foo', array('bar', array('foo', array('bar', 'foo')), 'foo'))),
+            ['[foo, [bar, [foo, [bar, foo]], foo]]', ['foo', ['bar', ['foo', ['bar', 'foo']], 'foo']]],
 
-            array('[foo, {bar: foo, foo: [foo, {bar: foo}]}, [foo, {bar: foo}]]', array('foo', array('bar' => 'foo', 'foo' => array('foo', array('bar' => 'foo'))), array('foo', array('bar' => 'foo')))),
+            ['[foo, {bar: foo, foo: [foo, {bar: foo}]}, [foo, {bar: foo}]]', ['foo', ['bar' => 'foo', 'foo' => ['foo', ['bar' => 'foo']]], ['foo', ['bar' => 'foo']]]],
 
-            array('[foo, bar: { foo: bar }]', array('foo', '1' => array('bar' => array('foo' => 'bar')))),
-            array('[foo, \'@foo.baz\', { \'%foo%\': \'foo is %foo%\', bar: \'%foo%\' }, true, \'@service_container\']', array('foo', '@foo.baz', array('%foo%' => 'foo is %foo%', 'bar' => '%foo%'), true, '@service_container')),
-        );
+            ['[foo, bar: { foo: bar }]', ['foo', '1' => ['bar' => ['foo' => 'bar']]]],
+            ['[foo, \'@foo.baz\', { \'%foo%\': \'foo is %foo%\', bar: \'%foo%\' }, true, \'@service_container\']', ['foo', '@foo.baz', ['%foo%' => 'foo is %foo%', 'bar' => '%foo%'], true, '@service_container']],
+        ];
     }
 
     public function getTestsForParseWithMapObjects()
     {
-        return array(
-            array('', ''),
-            array('null', null),
-            array('false', false),
-            array('true', true),
-            array('12', 12),
-            array('-12', -12),
-            array('"quoted string"', 'quoted string'),
-            array("'quoted string'", 'quoted string'),
-            array('12.30e+02', 12.30e+02),
-            array('0x4D2', 0x4D2),
-            array('02333', 02333),
-            array('.Inf', -log(0)),
-            array('-.Inf', log(0)),
-            array("'686e444'", '686e444'),
-            array('686e444', 646e444),
-            array('123456789123456789123456789123456789', '123456789123456789123456789123456789'),
-            array('"foo\r\nbar"', "foo\r\nbar"),
-            array("'foo#bar'", 'foo#bar'),
-            array("'foo # bar'", 'foo # bar'),
-            array("'#cfcfcf'", '#cfcfcf'),
-            array('::form_base.html.twig', '::form_base.html.twig'),
+        return [
+            ['', ''],
+            ['null', null],
+            ['false', false],
+            ['true', true],
+            ['12', 12],
+            ['-12', -12],
+            ['"quoted string"', 'quoted string'],
+            ["'quoted string'", 'quoted string'],
+            ['12.30e+02', 12.30e+02],
+            ['0x4D2', 0x4D2],
+            ['02333', 02333],
+            ['.Inf', -log(0)],
+            ['-.Inf', log(0)],
+            ["'686e444'", '686e444'],
+            ['686e444', 646e444],
+            ['123456789123456789123456789123456789', '123456789123456789123456789123456789'],
+            ['"foo\r\nbar"', "foo\r\nbar"],
+            ["'foo#bar'", 'foo#bar'],
+            ["'foo # bar'", 'foo # bar'],
+            ["'#cfcfcf'", '#cfcfcf'],
+            ['::form_base.html.twig', '::form_base.html.twig'],
 
-            array('2007-10-30', gmmktime(0, 0, 0, 10, 30, 2007)),
-            array('2007-10-30T02:59:43Z', gmmktime(2, 59, 43, 10, 30, 2007)),
-            array('2007-10-30 02:59:43 Z', gmmktime(2, 59, 43, 10, 30, 2007)),
-            array('1960-10-30 02:59:43 Z', gmmktime(2, 59, 43, 10, 30, 1960)),
-            array('1730-10-30T02:59:43Z', gmmktime(2, 59, 43, 10, 30, 1730)),
+            ['2007-10-30', gmmktime(0, 0, 0, 10, 30, 2007)],
+            ['2007-10-30T02:59:43Z', gmmktime(2, 59, 43, 10, 30, 2007)],
+            ['2007-10-30 02:59:43 Z', gmmktime(2, 59, 43, 10, 30, 2007)],
+            ['1960-10-30 02:59:43 Z', gmmktime(2, 59, 43, 10, 30, 1960)],
+            ['1730-10-30T02:59:43Z', gmmktime(2, 59, 43, 10, 30, 1730)],
 
-            array('"a \\"string\\" with \'quoted strings inside\'"', 'a "string" with \'quoted strings inside\''),
-            array("'a \"string\" with ''quoted strings inside'''", 'a "string" with \'quoted strings inside\''),
+            ['"a \\"string\\" with \'quoted strings inside\'"', 'a "string" with \'quoted strings inside\''],
+            ["'a \"string\" with ''quoted strings inside'''", 'a "string" with \'quoted strings inside\''],
 
             // sequences
             // urls are no key value mapping. see #3609. Valid yaml "key: value" mappings require a space after the colon
-            array('[foo, http://urls.are/no/mappings, false, null, 12]', array('foo', 'http://urls.are/no/mappings', false, null, 12)),
-            array('[  foo  ,   bar , false  ,  null     ,  12  ]', array('foo', 'bar', false, null, 12)),
-            array('[\'foo,bar\', \'foo bar\']', array('foo,bar', 'foo bar')),
+            ['[foo, http://urls.are/no/mappings, false, null, 12]', ['foo', 'http://urls.are/no/mappings', false, null, 12]],
+            ['[  foo  ,   bar , false  ,  null     ,  12  ]', ['foo', 'bar', false, null, 12]],
+            ['[\'foo,bar\', \'foo bar\']', ['foo,bar', 'foo bar']],
 
             // mappings
-            array('{foo: bar,bar: foo,"false": false,"null": null,integer: 12}', (object) array('foo' => 'bar', 'bar' => 'foo', 'false' => false, 'null' => null, 'integer' => 12), Yaml::PARSE_OBJECT_FOR_MAP),
-            array('{ foo  : bar, bar : foo,  "false"  :   false,  "null"  :   null,  integer :  12  }', (object) array('foo' => 'bar', 'bar' => 'foo', 'false' => false, 'null' => null, 'integer' => 12), Yaml::PARSE_OBJECT_FOR_MAP),
-            array('{foo: \'bar\', bar: \'foo: bar\'}', (object) array('foo' => 'bar', 'bar' => 'foo: bar')),
-            array('{\'foo\': \'bar\', "bar": \'foo: bar\'}', (object) array('foo' => 'bar', 'bar' => 'foo: bar')),
-            array('{\'foo\'\'\': \'bar\', "bar\"": \'foo: bar\'}', (object) array('foo\'' => 'bar', 'bar"' => 'foo: bar')),
-            array('{\'foo: \': \'bar\', "bar: ": \'foo: bar\'}', (object) array('foo: ' => 'bar', 'bar: ' => 'foo: bar')),
-            array('{"foo:bar": "baz"}', (object) array('foo:bar' => 'baz')),
-            array('{"foo":"bar"}', (object) array('foo' => 'bar')),
+            ['{foo: bar,bar: foo,"false": false,"null": null,integer: 12}', (object) ['foo' => 'bar', 'bar' => 'foo', 'false' => false, 'null' => null, 'integer' => 12], Yaml::PARSE_OBJECT_FOR_MAP],
+            ['{ foo  : bar, bar : foo,  "false"  :   false,  "null"  :   null,  integer :  12  }', (object) ['foo' => 'bar', 'bar' => 'foo', 'false' => false, 'null' => null, 'integer' => 12], Yaml::PARSE_OBJECT_FOR_MAP],
+            ['{foo: \'bar\', bar: \'foo: bar\'}', (object) ['foo' => 'bar', 'bar' => 'foo: bar']],
+            ['{\'foo\': \'bar\', "bar": \'foo: bar\'}', (object) ['foo' => 'bar', 'bar' => 'foo: bar']],
+            ['{\'foo\'\'\': \'bar\', "bar\"": \'foo: bar\'}', (object) ['foo\'' => 'bar', 'bar"' => 'foo: bar']],
+            ['{\'foo: \': \'bar\', "bar: ": \'foo: bar\'}', (object) ['foo: ' => 'bar', 'bar: ' => 'foo: bar']],
+            ['{"foo:bar": "baz"}', (object) ['foo:bar' => 'baz']],
+            ['{"foo":"bar"}', (object) ['foo' => 'bar']],
 
             // nested sequences and mappings
-            array('[foo, [bar, foo]]', array('foo', array('bar', 'foo'))),
-            array('[foo, {bar: foo}]', array('foo', (object) array('bar' => 'foo'))),
-            array('{ foo: {bar: foo} }', (object) array('foo' => (object) array('bar' => 'foo'))),
-            array('{ foo: [bar, foo] }', (object) array('foo' => array('bar', 'foo'))),
+            ['[foo, [bar, foo]]', ['foo', ['bar', 'foo']]],
+            ['[foo, {bar: foo}]', ['foo', (object) ['bar' => 'foo']]],
+            ['{ foo: {bar: foo} }', (object) ['foo' => (object) ['bar' => 'foo']]],
+            ['{ foo: [bar, foo] }', (object) ['foo' => ['bar', 'foo']]],
 
-            array('[  foo, [  bar, foo  ]  ]', array('foo', array('bar', 'foo'))),
+            ['[  foo, [  bar, foo  ]  ]', ['foo', ['bar', 'foo']]],
 
-            array('[{ foo: {bar: foo} }]', array((object) array('foo' => (object) array('bar' => 'foo')))),
+            ['[{ foo: {bar: foo} }]', [(object) ['foo' => (object) ['bar' => 'foo']]]],
 
-            array('[foo, [bar, [foo, [bar, foo]], foo]]', array('foo', array('bar', array('foo', array('bar', 'foo')), 'foo'))),
+            ['[foo, [bar, [foo, [bar, foo]], foo]]', ['foo', ['bar', ['foo', ['bar', 'foo']], 'foo']]],
 
-            array('[foo, {bar: foo, foo: [foo, {bar: foo}]}, [foo, {bar: foo}]]', array('foo', (object) array('bar' => 'foo', 'foo' => array('foo', (object) array('bar' => 'foo'))), array('foo', (object) array('bar' => 'foo')))),
+            ['[foo, {bar: foo, foo: [foo, {bar: foo}]}, [foo, {bar: foo}]]', ['foo', (object) ['bar' => 'foo', 'foo' => ['foo', (object) ['bar' => 'foo']]], ['foo', (object) ['bar' => 'foo']]]],
 
-            array('[foo, bar: { foo: bar }]', array('foo', '1' => (object) array('bar' => (object) array('foo' => 'bar')))),
-            array('[foo, \'@foo.baz\', { \'%foo%\': \'foo is %foo%\', bar: \'%foo%\' }, true, \'@service_container\']', array('foo', '@foo.baz', (object) array('%foo%' => 'foo is %foo%', 'bar' => '%foo%'), true, '@service_container')),
+            ['[foo, bar: { foo: bar }]', ['foo', '1' => (object) ['bar' => (object) ['foo' => 'bar']]]],
+            ['[foo, \'@foo.baz\', { \'%foo%\': \'foo is %foo%\', bar: \'%foo%\' }, true, \'@service_container\']', ['foo', '@foo.baz', (object) ['%foo%' => 'foo is %foo%', 'bar' => '%foo%'], true, '@service_container']],
 
-            array('{}', new \stdClass()),
-            array('{ foo  : bar, bar : {}  }', (object) array('foo' => 'bar', 'bar' => new \stdClass())),
-            array('{ foo  : [], bar : {}  }', (object) array('foo' => array(), 'bar' => new \stdClass())),
-            array('{foo: \'bar\', bar: {} }', (object) array('foo' => 'bar', 'bar' => new \stdClass())),
-            array('{\'foo\': \'bar\', "bar": {}}', (object) array('foo' => 'bar', 'bar' => new \stdClass())),
-            array('{\'foo\': \'bar\', "bar": \'{}\'}', (object) array('foo' => 'bar', 'bar' => '{}')),
+            ['{}', new \stdClass()],
+            ['{ foo  : bar, bar : {}  }', (object) ['foo' => 'bar', 'bar' => new \stdClass()]],
+            ['{ foo  : [], bar : {}  }', (object) ['foo' => [], 'bar' => new \stdClass()]],
+            ['{foo: \'bar\', bar: {} }', (object) ['foo' => 'bar', 'bar' => new \stdClass()]],
+            ['{\'foo\': \'bar\', "bar": {}}', (object) ['foo' => 'bar', 'bar' => new \stdClass()]],
+            ['{\'foo\': \'bar\', "bar": \'{}\'}', (object) ['foo' => 'bar', 'bar' => '{}']],
 
-            array('[foo, [{}, {}]]', array('foo', array(new \stdClass(), new \stdClass()))),
-            array('[foo, [[], {}]]', array('foo', array(array(), new \stdClass()))),
-            array('[foo, [[{}, {}], {}]]', array('foo', array(array(new \stdClass(), new \stdClass()), new \stdClass()))),
-            array('[foo, {bar: {}}]', array('foo', '1' => (object) array('bar' => new \stdClass()))),
-        );
+            ['[foo, [{}, {}]]', ['foo', [new \stdClass(), new \stdClass()]]],
+            ['[foo, [[], {}]]', ['foo', [[], new \stdClass()]]],
+            ['[foo, [[{}, {}], {}]]', ['foo', [[new \stdClass(), new \stdClass()], new \stdClass()]]],
+            ['[foo, {bar: {}}]', ['foo', '1' => (object) ['bar' => new \stdClass()]]],
+        ];
     }
 
     public function getTestsForDump()
     {
-        return array(
-            array('null', null),
-            array('false', false),
-            array('true', true),
-            array('12', 12),
-            array("'1_2'", '1_2'),
-            array('_12', '_12'),
-            array("'12_'", '12_'),
-            array("'quoted string'", 'quoted string'),
-            array('!!float 1230', 12.30e+02),
-            array('1234', 0x4D2),
-            array('1243', 02333),
-            array("'0x_4_D_2_'", '0x_4_D_2_'),
-            array("'0_2_3_3_3'", '0_2_3_3_3'),
-            array('.Inf', -log(0)),
-            array('-.Inf', log(0)),
-            array("'686e444'", '686e444'),
-            array('"foo\r\nbar"', "foo\r\nbar"),
-            array("'foo#bar'", 'foo#bar'),
-            array("'foo # bar'", 'foo # bar'),
-            array("'#cfcfcf'", '#cfcfcf'),
+        return [
+            ['null', null],
+            ['false', false],
+            ['true', true],
+            ['12', 12],
+            ["'1_2'", '1_2'],
+            ['_12', '_12'],
+            ["'12_'", '12_'],
+            ["'quoted string'", 'quoted string'],
+            ['!!float 1230', 12.30e+02],
+            ['1234', 0x4D2],
+            ['1243', 02333],
+            ["'0x_4_D_2_'", '0x_4_D_2_'],
+            ["'0_2_3_3_3'", '0_2_3_3_3'],
+            ['.Inf', -log(0)],
+            ['-.Inf', log(0)],
+            ["'686e444'", '686e444'],
+            ['"foo\r\nbar"', "foo\r\nbar"],
+            ["'foo#bar'", 'foo#bar'],
+            ["'foo # bar'", 'foo # bar'],
+            ["'#cfcfcf'", '#cfcfcf'],
 
-            array("'a \"string\" with ''quoted strings inside'''", 'a "string" with \'quoted strings inside\''),
+            ["'a \"string\" with ''quoted strings inside'''", 'a "string" with \'quoted strings inside\''],
 
-            array("'-dash'", '-dash'),
-            array("'-'", '-'),
+            ["'-dash'", '-dash'],
+            ["'-'", '-'],
 
             // Pre-YAML-1.2 booleans
-            array("'y'", 'y'),
-            array("'n'", 'n'),
-            array("'yes'", 'yes'),
-            array("'no'", 'no'),
-            array("'on'", 'on'),
-            array("'off'", 'off'),
+            ["'y'", 'y'],
+            ["'n'", 'n'],
+            ["'yes'", 'yes'],
+            ["'no'", 'no'],
+            ["'on'", 'on'],
+            ["'off'", 'off'],
 
             // sequences
-            array('[foo, bar, false, null, 12]', array('foo', 'bar', false, null, 12)),
-            array('[\'foo,bar\', \'foo bar\']', array('foo,bar', 'foo bar')),
+            ['[foo, bar, false, null, 12]', ['foo', 'bar', false, null, 12]],
+            ['[\'foo,bar\', \'foo bar\']', ['foo,bar', 'foo bar']],
 
             // mappings
-            array('{ foo: bar, bar: foo, \'false\': false, \'null\': null, integer: 12 }', array('foo' => 'bar', 'bar' => 'foo', 'false' => false, 'null' => null, 'integer' => 12)),
-            array('{ foo: bar, bar: \'foo: bar\' }', array('foo' => 'bar', 'bar' => 'foo: bar')),
+            ['{ foo: bar, bar: foo, \'false\': false, \'null\': null, integer: 12 }', ['foo' => 'bar', 'bar' => 'foo', 'false' => false, 'null' => null, 'integer' => 12]],
+            ['{ foo: bar, bar: \'foo: bar\' }', ['foo' => 'bar', 'bar' => 'foo: bar']],
 
             // nested sequences and mappings
-            array('[foo, [bar, foo]]', array('foo', array('bar', 'foo'))),
+            ['[foo, [bar, foo]]', ['foo', ['bar', 'foo']]],
 
-            array('[foo, [bar, [foo, [bar, foo]], foo]]', array('foo', array('bar', array('foo', array('bar', 'foo')), 'foo'))),
+            ['[foo, [bar, [foo, [bar, foo]], foo]]', ['foo', ['bar', ['foo', ['bar', 'foo']], 'foo']]],
 
-            array('{ foo: { bar: foo } }', array('foo' => array('bar' => 'foo'))),
+            ['{ foo: { bar: foo } }', ['foo' => ['bar' => 'foo']]],
 
-            array('[foo, { bar: foo }]', array('foo', array('bar' => 'foo'))),
+            ['[foo, { bar: foo }]', ['foo', ['bar' => 'foo']]],
 
-            array('[foo, { bar: foo, foo: [foo, { bar: foo }] }, [foo, { bar: foo }]]', array('foo', array('bar' => 'foo', 'foo' => array('foo', array('bar' => 'foo'))), array('foo', array('bar' => 'foo')))),
+            ['[foo, { bar: foo, foo: [foo, { bar: foo }] }, [foo, { bar: foo }]]', ['foo', ['bar' => 'foo', 'foo' => ['foo', ['bar' => 'foo']]], ['foo', ['bar' => 'foo']]]],
 
-            array('[foo, \'@foo.baz\', { \'%foo%\': \'foo is %foo%\', bar: \'%foo%\' }, true, \'@service_container\']', array('foo', '@foo.baz', array('%foo%' => 'foo is %foo%', 'bar' => '%foo%'), true, '@service_container')),
+            ['[foo, \'@foo.baz\', { \'%foo%\': \'foo is %foo%\', bar: \'%foo%\' }, true, \'@service_container\']', ['foo', '@foo.baz', ['%foo%' => 'foo is %foo%', 'bar' => '%foo%'], true, '@service_container']],
 
-            array('{ foo: { bar: { 1: 2, baz: 3 } } }', array('foo' => array('bar' => array(1 => 2, 'baz' => 3)))),
-        );
+            ['{ foo: { bar: { 1: 2, baz: 3 } } }', ['foo' => ['bar' => [1 => 2, 'baz' => 3]]]],
+        ];
     }
 
     /**
@@ -556,12 +556,12 @@ class InlineTest extends TestCase
 
     public function getTimestampTests()
     {
-        return array(
-            'canonical' => array('2001-12-15T02:59:43.1Z', 2001, 12, 15, 2, 59, 43.1, '+0000'),
-            'ISO-8601' => array('2001-12-15t21:59:43.10-05:00', 2001, 12, 16, 2, 59, 43.1, '-0500'),
-            'spaced' => array('2001-12-15 21:59:43.10 -5', 2001, 12, 16, 2, 59, 43.1, '-0500'),
-            'date' => array('2001-12-15', 2001, 12, 15, 0, 0, 0, '+0000'),
-        );
+        return [
+            'canonical' => ['2001-12-15T02:59:43.1Z', 2001, 12, 15, 2, 59, 43.1, '+0000'],
+            'ISO-8601' => ['2001-12-15t21:59:43.10-05:00', 2001, 12, 16, 2, 59, 43.1, '-0500'],
+            'spaced' => ['2001-12-15 21:59:43.10 -5', 2001, 12, 16, 2, 59, 43.1, '-0500'],
+            'date' => ['2001-12-15', 2001, 12, 15, 0, 0, 0, '+0000'],
+        ];
     }
 
     /**
@@ -574,7 +574,7 @@ class InlineTest extends TestCase
         $expected->setDate($year, $month, $day);
         $expected->setTime($hour, $minute, $second, 1000000 * ($second - (int) $second));
 
-        $expectedNested = array('nested' => array($expected));
+        $expectedNested = ['nested' => [$expected]];
         $yamlNested = "{nested: [$yaml]}";
 
         $this->assertEquals($expectedNested, Inline::parse($yamlNested, Yaml::PARSE_DATETIME));
@@ -590,13 +590,13 @@ class InlineTest extends TestCase
 
     public function getDateTimeDumpTests()
     {
-        $tests = array();
+        $tests = [];
 
         $dateTime = new \DateTime('2001-12-15 21:59:43', new \DateTimeZone('UTC'));
-        $tests['date-time-utc'] = array($dateTime, '2001-12-15T21:59:43+00:00');
+        $tests['date-time-utc'] = [$dateTime, '2001-12-15T21:59:43+00:00'];
 
         $dateTime = new \DateTimeImmutable('2001-07-15 21:59:43', new \DateTimeZone('Europe/Berlin'));
-        $tests['immutable-date-time-europe-berlin'] = array($dateTime, '2001-07-15T21:59:43+02:00');
+        $tests['immutable-date-time-europe-berlin'] = [$dateTime, '2001-07-15T21:59:43+02:00'];
 
         return $tests;
     }
@@ -611,11 +611,11 @@ class InlineTest extends TestCase
 
     public function getBinaryData()
     {
-        return array(
-            'enclosed with double quotes' => array('!!binary "SGVsbG8gd29ybGQ="'),
-            'enclosed with single quotes' => array("!!binary 'SGVsbG8gd29ybGQ='"),
-            'containing spaces' => array('!!binary  "SGVs bG8gd 29ybGQ="'),
-        );
+        return [
+            'enclosed with double quotes' => ['!!binary "SGVsbG8gd29ybGQ="'],
+            'enclosed with single quotes' => ["!!binary 'SGVsbG8gd29ybGQ='"],
+            'containing spaces' => ['!!binary  "SGVs bG8gd 29ybGQ="'],
+        ];
     }
 
     /**
@@ -635,12 +635,12 @@ class InlineTest extends TestCase
 
     public function getInvalidBinaryData()
     {
-        return array(
-            'length not a multiple of four' => array('!!binary "SGVsbG8d29ybGQ="', '/The normalized base64 encoded data \(data without whitespace characters\) length must be a multiple of four \(\d+ bytes given\)/'),
-            'invalid characters' => array('!!binary "SGVsbG8#d29ybGQ="', '/The base64 encoded data \(.*\) contains invalid characters/'),
-            'too many equals characters' => array('!!binary "SGVsbG8gd29yb==="', '/The base64 encoded data \(.*\) contains invalid characters/'),
-            'misplaced equals character' => array('!!binary "SGVsbG8gd29ybG=Q"', '/The base64 encoded data \(.*\) contains invalid characters/'),
-        );
+        return [
+            'length not a multiple of four' => ['!!binary "SGVsbG8d29ybGQ="', '/The normalized base64 encoded data \(data without whitespace characters\) length must be a multiple of four \(\d+ bytes given\)/'],
+            'invalid characters' => ['!!binary "SGVsbG8#d29ybGQ="', '/The base64 encoded data \(.*\) contains invalid characters/'],
+            'too many equals characters' => ['!!binary "SGVsbG8gd29yb==="', '/The base64 encoded data \(.*\) contains invalid characters/'],
+            'misplaced equals character' => ['!!binary "SGVsbG8gd29ybG=Q"', '/The base64 encoded data \(.*\) contains invalid characters/'],
+        ];
     }
 
     /**
@@ -656,7 +656,7 @@ class InlineTest extends TestCase
     {
         $longStringWithQuotes = str_repeat("x\r\n\\\"x\"x", 1000);
 
-        $yamlString = Inline::dump(array('longStringWithQuotes' => $longStringWithQuotes));
+        $yamlString = Inline::dump(['longStringWithQuotes' => $longStringWithQuotes]);
         $arrayFromYaml = Inline::parse($yamlString);
 
         $this->assertEquals($longStringWithQuotes, $arrayFromYaml['longStringWithQuotes']);
@@ -681,15 +681,15 @@ class InlineTest extends TestCase
 
     public function getTestsForNullValues()
     {
-        return array(
-            'null before closing curly brace' => array('{foo:}', array('foo' => null)),
-            'null before comma' => array('{foo:, bar: baz}', array('foo' => null, 'bar' => 'baz')),
-        );
+        return [
+            'null before closing curly brace' => ['{foo:}', ['foo' => null]],
+            'null before comma' => ['{foo:, bar: baz}', ['foo' => null, 'bar' => 'baz']],
+        ];
     }
 
     public function testTheEmptyStringIsAValidMappingKey()
     {
-        $this->assertSame(array('' => 'foo'), Inline::parse('{ "": foo }'));
+        $this->assertSame(['' => 'foo'], Inline::parse('{ "": foo }'));
     }
 
     /**
@@ -705,12 +705,12 @@ class InlineTest extends TestCase
 
     public function getNotPhpCompatibleMappingKeyData()
     {
-        return array(
-            'boolean-true' => array('{true: "foo"}', array('true' => 'foo')),
-            'boolean-false' => array('{false: "foo"}', array('false' => 'foo')),
-            'null' => array('{null: "foo"}', array('null' => 'foo')),
-            'float' => array('{0.25: "foo"}', array('0.25' => 'foo')),
-        );
+        return [
+            'boolean-true' => ['{true: "foo"}', ['true' => 'foo']],
+            'boolean-false' => ['{false: "foo"}', ['false' => 'foo']],
+            'null' => ['{null: "foo"}', ['null' => 'foo']],
+            'float' => ['{0.25: "foo"}', ['0.25' => 'foo']],
+        ];
     }
 
     public function testTagWithoutValueInSequence()
