@@ -55,12 +55,12 @@ class ParserTest extends TestCase
     public function testTabsInYaml()
     {
         // test tabs in YAML
-        $yamls = array(
+        $yamls = [
             "foo:\n	bar",
             "foo:\n 	bar",
             "foo:\n	 bar",
             "foo:\n 	 bar",
-        );
+        ];
 
         foreach ($yamls as $yaml) {
             try {
@@ -87,7 +87,7 @@ EOF;
 
     public function getBlockChompingTests()
     {
-        $tests = array();
+        $tests = [];
 
         $yaml = <<<'EOF'
 foo: |-
@@ -98,11 +98,11 @@ bar: |-
     two
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one\ntwo",
             'bar' => "one\ntwo",
-        );
-        $tests['Literal block chomping strip with single trailing newline'] = array($expected, $yaml);
+        ];
+        $tests['Literal block chomping strip with single trailing newline'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: |-
@@ -115,19 +115,19 @@ bar: |-
 
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one\ntwo",
             'bar' => "one\ntwo",
-        );
-        $tests['Literal block chomping strip with multiple trailing newlines'] = array($expected, $yaml);
+        ];
+        $tests['Literal block chomping strip with multiple trailing newlines'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 {}
 
 
 EOF;
-        $expected = array();
-        $tests['Literal block chomping strip with multiple trailing newlines after a 1-liner'] = array($expected, $yaml);
+        $expected = [];
+        $tests['Literal block chomping strip with multiple trailing newlines after a 1-liner'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: |-
@@ -137,11 +137,11 @@ bar: |-
     one
     two
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one\ntwo",
             'bar' => "one\ntwo",
-        );
-        $tests['Literal block chomping strip without trailing newline'] = array($expected, $yaml);
+        ];
+        $tests['Literal block chomping strip without trailing newline'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: |
@@ -152,11 +152,11 @@ bar: |
     two
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one\ntwo\n",
             'bar' => "one\ntwo\n",
-        );
-        $tests['Literal block chomping clip with single trailing newline'] = array($expected, $yaml);
+        ];
+        $tests['Literal block chomping clip with single trailing newline'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: |
@@ -169,11 +169,11 @@ bar: |
 
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one\ntwo\n",
             'bar' => "one\ntwo\n",
-        );
-        $tests['Literal block chomping clip with multiple trailing newlines'] = array($expected, $yaml);
+        ];
+        $tests['Literal block chomping clip with multiple trailing newlines'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo:
@@ -182,14 +182,14 @@ foo:
 
     two
 EOF;
-        $expected = array(
-            'foo' => array(
-                array(
+        $expected = [
+            'foo' => [
+                [
                     'bar' => "one\n\ntwo",
-                ),
-            ),
-        );
-        $tests['Literal block chomping clip with embedded blank line inside unindented collection'] = array($expected, $yaml);
+                ],
+            ],
+        ];
+        $tests['Literal block chomping clip with embedded blank line inside unindented collection'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: |
@@ -199,11 +199,11 @@ bar: |
     one
     two
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one\ntwo\n",
             'bar' => "one\ntwo",
-        );
-        $tests['Literal block chomping clip without trailing newline'] = array($expected, $yaml);
+        ];
+        $tests['Literal block chomping clip without trailing newline'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: |+
@@ -214,11 +214,11 @@ bar: |+
     two
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one\ntwo\n",
             'bar' => "one\ntwo\n",
-        );
-        $tests['Literal block chomping keep with single trailing newline'] = array($expected, $yaml);
+        ];
+        $tests['Literal block chomping keep with single trailing newline'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: |+
@@ -231,11 +231,11 @@ bar: |+
 
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one\ntwo\n\n",
             'bar' => "one\ntwo\n\n",
-        );
-        $tests['Literal block chomping keep with multiple trailing newlines'] = array($expected, $yaml);
+        ];
+        $tests['Literal block chomping keep with multiple trailing newlines'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: |+
@@ -245,11 +245,11 @@ bar: |+
     one
     two
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one\ntwo\n",
             'bar' => "one\ntwo",
-        );
-        $tests['Literal block chomping keep without trailing newline'] = array($expected, $yaml);
+        ];
+        $tests['Literal block chomping keep without trailing newline'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: >-
@@ -260,11 +260,11 @@ bar: >-
     two
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => 'one two',
             'bar' => 'one two',
-        );
-        $tests['Folded block chomping strip with single trailing newline'] = array($expected, $yaml);
+        ];
+        $tests['Folded block chomping strip with single trailing newline'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: >-
@@ -277,11 +277,11 @@ bar: >-
 
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => 'one two',
             'bar' => 'one two',
-        );
-        $tests['Folded block chomping strip with multiple trailing newlines'] = array($expected, $yaml);
+        ];
+        $tests['Folded block chomping strip with multiple trailing newlines'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: >-
@@ -291,11 +291,11 @@ bar: >-
     one
     two
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => 'one two',
             'bar' => 'one two',
-        );
-        $tests['Folded block chomping strip without trailing newline'] = array($expected, $yaml);
+        ];
+        $tests['Folded block chomping strip without trailing newline'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: >
@@ -306,11 +306,11 @@ bar: >
     two
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one two\n",
             'bar' => "one two\n",
-        );
-        $tests['Folded block chomping clip with single trailing newline'] = array($expected, $yaml);
+        ];
+        $tests['Folded block chomping clip with single trailing newline'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: >
@@ -323,11 +323,11 @@ bar: >
 
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one two\n",
             'bar' => "one two\n",
-        );
-        $tests['Folded block chomping clip with multiple trailing newlines'] = array($expected, $yaml);
+        ];
+        $tests['Folded block chomping clip with multiple trailing newlines'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: >
@@ -337,11 +337,11 @@ bar: >
     one
     two
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one two\n",
             'bar' => 'one two',
-        );
-        $tests['Folded block chomping clip without trailing newline'] = array($expected, $yaml);
+        ];
+        $tests['Folded block chomping clip without trailing newline'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: >+
@@ -352,11 +352,11 @@ bar: >+
     two
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one two\n",
             'bar' => "one two\n",
-        );
-        $tests['Folded block chomping keep with single trailing newline'] = array($expected, $yaml);
+        ];
+        $tests['Folded block chomping keep with single trailing newline'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: >+
@@ -369,11 +369,11 @@ bar: >+
 
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one two\n\n",
             'bar' => "one two\n\n",
-        );
-        $tests['Folded block chomping keep with multiple trailing newlines'] = array($expected, $yaml);
+        ];
+        $tests['Folded block chomping keep with multiple trailing newlines'] = [$expected, $yaml];
 
         $yaml = <<<'EOF'
 foo: >+
@@ -383,11 +383,11 @@ bar: >+
     one
     two
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "one two\n",
             'bar' => 'one two',
-        );
-        $tests['Folded block chomping keep without trailing newline'] = array($expected, $yaml);
+        ];
+        $tests['Folded block chomping keep without trailing newline'] = [$expected, $yaml];
 
         return $tests;
     }
@@ -414,9 +414,9 @@ foo: |-
     bar
 
 EOF;
-        $expected = array(
+        $expected = [
             'foo' => "\n\nbar",
-        );
+        ];
 
         $this->assertSame($expected, $this->parser->parse($yaml));
     }
@@ -427,7 +427,7 @@ EOF;
 foo: !php/object O:30:"Symfony\Component\Yaml\Tests\B":1:{s:1:"b";s:3:"foo";}
 bar: 1
 EOF;
-        $this->assertEquals(array('foo' => new B(), 'bar' => 1), $this->parser->parse($input, Yaml::PARSE_OBJECT), '->parse() is able to parse objects');
+        $this->assertEquals(['foo' => new B(), 'bar' => 1], $this->parser->parse($input, Yaml::PARSE_OBJECT), '->parse() is able to parse objects');
     }
 
     public function testObjectSupportDisabledButNoExceptions()
@@ -436,7 +436,7 @@ EOF;
 foo: !php/object O:30:"Symfony\Tests\Component\Yaml\B":1:{s:1:"b";s:3:"foo";}
 bar: 1
 EOF;
-        $this->assertEquals(array('foo' => null, 'bar' => 1), $this->parser->parse($input), '->parse() does not parse objects');
+        $this->assertEquals(['foo' => null, 'bar' => 1], $this->parser->parse($input), '->parse() does not parse objects');
     }
 
     /**
@@ -451,7 +451,7 @@ EOF;
 
     public function getObjectForMapTests()
     {
-        $tests = array();
+        $tests = [];
 
         $yaml = <<<'EOF'
 foo:
@@ -459,20 +459,20 @@ foo:
 EOF;
         $expected = new \stdClass();
         $expected->foo = new \stdClass();
-        $expected->foo->fiz = array('cat');
-        $tests['mapping'] = array($yaml, $expected);
+        $expected->foo->fiz = ['cat'];
+        $tests['mapping'] = [$yaml, $expected];
 
         $yaml = '{ "foo": "bar", "fiz": "cat" }';
         $expected = new \stdClass();
         $expected->foo = 'bar';
         $expected->fiz = 'cat';
-        $tests['inline-mapping'] = array($yaml, $expected);
+        $tests['inline-mapping'] = [$yaml, $expected];
 
         $yaml = "foo: bar\nbaz: foobar";
         $expected = new \stdClass();
         $expected->foo = 'bar';
         $expected->baz = 'foobar';
-        $tests['object-for-map-is-applied-after-parsing'] = array($yaml, $expected);
+        $tests['object-for-map-is-applied-after-parsing'] = [$yaml, $expected];
 
         $yaml = <<<'EOT'
 array:
@@ -480,12 +480,12 @@ array:
   - key: two
 EOT;
         $expected = new \stdClass();
-        $expected->array = array();
+        $expected->array = [];
         $expected->array[0] = new \stdClass();
         $expected->array[0]->key = 'one';
         $expected->array[1] = new \stdClass();
         $expected->array[1]->key = 'two';
-        $tests['nest-map-and-sequence'] = array($yaml, $expected);
+        $tests['nest-map-and-sequence'] = [$yaml, $expected];
 
         $yaml = <<<'YAML'
 map:
@@ -496,7 +496,7 @@ YAML;
         $expected->map = new \stdClass();
         $expected->map->{1} = 'one';
         $expected->map->{2} = 'two';
-        $tests['numeric-keys'] = array($yaml, $expected);
+        $tests['numeric-keys'] = [$yaml, $expected];
 
         $yaml = <<<'YAML'
 map:
@@ -507,7 +507,7 @@ YAML;
         $expected->map = new \stdClass();
         $expected->map->{0} = 'one';
         $expected->map->{1} = 'two';
-        $tests['zero-indexed-numeric-keys'] = array($yaml, $expected);
+        $tests['zero-indexed-numeric-keys'] = [$yaml, $expected];
 
         return $tests;
     }
@@ -529,9 +529,9 @@ EOF;
     {
         $yaml = "items:  \n  foo: bar";
 
-        $expected = array(
-            'items' => array('foo' => 'bar'),
-        );
+        $expected = [
+            'items' => ['foo' => 'bar'],
+        ];
 
         $this->assertSame($expected, $this->parser->parse($yaml));
     }
@@ -541,11 +541,11 @@ EOF;
      */
     public function testNonUtf8Exception()
     {
-        $yamls = array(
+        $yamls = [
             iconv('UTF-8', 'ISO-8859-1', "foo: 'äöüß'"),
             iconv('UTF-8', 'ISO-8859-15', "euro: '€'"),
             iconv('UTF-8', 'CP1252', "cp1252: '©ÉÇáñ'"),
-        );
+        ];
 
         foreach ($yamls as $yaml) {
             try {
@@ -636,19 +636,19 @@ a:
 - foo
 d: e
 EOT;
-        $expected = array(
-            'a' => array(
-                array(
-                    'b' => array(
-                        array(
+        $expected = [
+            'a' => [
+                [
+                    'b' => [
+                        [
                             'bar' => 'baz',
-                        ),
-                    ),
-                ),
+                        ],
+                    ],
+                ],
                 'foo',
-            ),
+            ],
             'd' => 'e',
-        );
+        ];
 
         $this->assertSame($expected, $this->parser->parse($yaml));
     }
@@ -662,12 +662,12 @@ a:
 # comment
     d: e
 EOT;
-        $expected = array(
-            'a' => array(
-                'b' => array('c'),
+        $expected = [
+            'a' => [
+                'b' => ['c'],
                 'd' => 'e',
-            ),
-        );
+            ],
+        ];
 
         $this->assertSame($expected, $this->parser->parse($yaml));
     }
@@ -683,25 +683,25 @@ a:
         1.1
 # another comment
 EOT;
-        $expected = array(
-            'a' => array(
-                'b' => array(),
+        $expected = [
+            'a' => [
+                'b' => [],
                 'd' => 1.1,
-            ),
-        );
+            ],
+        ];
 
         $this->assertSame($expected, $this->parser->parse($yaml));
     }
 
     public function getParseExceptionNotAffectedMultiLineStringLastResortParsing()
     {
-        $tests = array();
+        $tests = [];
 
         $yaml = <<<'EOT'
 a
     b:
 EOT;
-        $tests['parse error on first line'] = array($yaml);
+        $tests['parse error on first line'] = [$yaml];
 
         $yaml = <<<'EOT'
 a
@@ -709,12 +709,12 @@ a
 b
     c:
 EOT;
-        $tests['parse error due to inconsistent indentation'] = array($yaml);
+        $tests['parse error due to inconsistent indentation'] = [$yaml];
 
         $yaml = <<<'EOT'
  &  *  !  |  >  '  "  %  @  ` #, { asd a;sdasd }-@^qw3
 EOT;
-        $tests['symfony/symfony/issues/22967#issuecomment-322067742'] = array($yaml);
+        $tests['symfony/symfony/issues/22967#issuecomment-322067742'] = [$yaml];
 
         return $tests;
     }
@@ -736,9 +736,9 @@ test:
   true
   yes you can
 EOT;
-        $expected = array(
+        $expected = [
             'test' => 'You can have things that don\'t look like strings here true yes you can',
-        );
+        ];
 
         $this->assertSame($expected, $this->parser->parse($yaml));
 
@@ -747,9 +747,9 @@ a:
     b
        c
 EOT;
-        $expected = array(
+        $expected = [
             'a' => 'b c',
-        );
+        ];
 
         $this->assertSame($expected, $this->parser->parse($yaml));
     }
@@ -805,11 +805,11 @@ parent:
     child: duplicate
     child: duplicate
 EOD;
-        $expected = array(
-            'parent' => array(
+        $expected = [
+            'parent' => [
                 'child' => 'first',
-            ),
-        );
+            ],
+        ];
         $this->assertSame($expected, Yaml::parse($input));
     }
 
@@ -823,11 +823,11 @@ EOD;
 parent: { child: first, child: duplicate }
 parent: { child: duplicate, child: duplicate }
 EOD;
-        $expected = array(
-            'parent' => array(
+        $expected = [
+            'parent' => [
                 'child' => 'first',
-            ),
-        );
+            ],
+        ];
         $this->assertSame($expected, Yaml::parse($input));
     }
 
@@ -844,30 +844,30 @@ EOD;
 
     public function getParseExceptionOnDuplicateData()
     {
-        $tests = array();
+        $tests = [];
 
         $yaml = <<<EOD
 parent: { child: first, child: duplicate }
 EOD;
-        $tests[] = array($yaml, 'child', 1);
+        $tests[] = [$yaml, 'child', 1];
 
         $yaml = <<<EOD
 parent:
   child: first,
   child: duplicate
 EOD;
-        $tests[] = array($yaml, 'child', 3);
+        $tests[] = [$yaml, 'child', 3];
 
         $yaml = <<<EOD
 parent: { child: foo }
 parent: { child: bar }
 EOD;
-        $tests[] = array($yaml, 'parent', 2);
+        $tests[] = [$yaml, 'parent', 2];
 
         $yaml = <<<EOD
 parent: { child_mapping: { value: bar},  child_mapping: { value: bar} }
 EOD;
-        $tests[] = array($yaml, 'child_mapping', 1);
+        $tests[] = [$yaml, 'child_mapping', 1];
 
         $yaml = <<<EOD
 parent:
@@ -876,12 +876,12 @@ parent:
   child_mapping:
     value: bar
 EOD;
-        $tests[] = array($yaml, 'child_mapping', 4);
+        $tests[] = [$yaml, 'child_mapping', 4];
 
         $yaml = <<<EOD
 parent: { child_sequence: ['key1', 'key2', 'key3'],  child_sequence: ['key1', 'key2', 'key3'] }
 EOD;
-        $tests[] = array($yaml, 'child_sequence', 1);
+        $tests[] = [$yaml, 'child_sequence', 1];
 
         $yaml = <<<EOD
 parent:
@@ -894,7 +894,7 @@ parent:
     - key2
     - key3
 EOD;
-        $tests[] = array($yaml, 'child_sequence', 6);
+        $tests[] = [$yaml, 'child_sequence', 6];
 
         return $tests;
     }
@@ -905,21 +905,21 @@ EOD;
 hash:
 EOF;
 
-        $this->assertEquals(array('hash' => null), Yaml::parse($input));
+        $this->assertEquals(['hash' => null], Yaml::parse($input));
     }
 
     public function testCommentAtTheRootIndent()
     {
-        $this->assertEquals(array(
-            'services' => array(
-                'app.foo_service' => array(
+        $this->assertEquals([
+            'services' => [
+                'app.foo_service' => [
                     'class' => 'Foo',
-                ),
-                'app/bar_service' => array(
+                ],
+                'app/bar_service' => [
                     'class' => 'Bar',
-                ),
-            ),
-        ), Yaml::parse(<<<'EOF'
+                ],
+            ],
+        ], Yaml::parse(<<<'EOF'
 # comment 1
 services:
 # comment 2
@@ -936,7 +936,7 @@ EOF
 
     public function testStringBlockWithComments()
     {
-        $this->assertEquals(array('content' => <<<'EOT'
+        $this->assertEquals(['content' => <<<'EOT'
 # comment 1
 header
 
@@ -947,7 +947,7 @@ header
 
 footer # comment3
 EOT
-        ), Yaml::parse(<<<'EOF'
+        ], Yaml::parse(<<<'EOF'
 content: |
     # comment 1
     header
@@ -964,7 +964,7 @@ EOF
 
     public function testFoldedStringBlockWithComments()
     {
-        $this->assertEquals(array(array('content' => <<<'EOT'
+        $this->assertEquals([['content' => <<<'EOT'
 # comment 1
 header
 
@@ -975,7 +975,7 @@ header
 
 footer # comment3
 EOT
-        )), Yaml::parse(<<<'EOF'
+        ]], Yaml::parse(<<<'EOF'
 -
     content: |
         # comment 1
@@ -993,7 +993,7 @@ EOF
 
     public function testNestedFoldedStringBlockWithComments()
     {
-        $this->assertEquals(array(array(
+        $this->assertEquals([[
             'title' => 'some title',
             'content' => <<<'EOT'
 # comment 1
@@ -1006,7 +1006,7 @@ header
 
 footer # comment3
 EOT
-        )), Yaml::parse(<<<'EOF'
+        ]], Yaml::parse(<<<'EOF'
 -
     title: some title
     content: |
@@ -1025,17 +1025,17 @@ EOF
 
     public function testReferenceResolvingInInlineStrings()
     {
-        $this->assertEquals(array(
+        $this->assertEquals([
             'var' => 'var-value',
             'scalar' => 'var-value',
-            'list' => array('var-value'),
-            'list_in_list' => array(array('var-value')),
-            'map_in_list' => array(array('key' => 'var-value')),
-            'embedded_mapping' => array(array('key' => 'var-value')),
-            'map' => array('key' => 'var-value'),
-            'list_in_map' => array('key' => array('var-value')),
-            'map_in_map' => array('foo' => array('bar' => 'var-value')),
-        ), Yaml::parse(<<<'EOF'
+            'list' => ['var-value'],
+            'list_in_list' => [['var-value']],
+            'map_in_list' => [['key' => 'var-value']],
+            'embedded_mapping' => [['key' => 'var-value']],
+            'map' => ['key' => 'var-value'],
+            'list_in_map' => ['key' => ['var-value']],
+            'map_in_map' => ['foo' => ['bar' => 'var-value']],
+        ], Yaml::parse(<<<'EOF'
 var:  &var var-value
 scalar: *var
 list: [ *var ]
@@ -1057,7 +1057,7 @@ EOF
 foo: 1
 bar: 2
 EOF;
-        $this->assertEquals(array('foo' => 1, 'bar' => 2), $this->parser->parse($yaml));
+        $this->assertEquals(['foo' => 1, 'bar' => 2], $this->parser->parse($yaml));
     }
 
     /**
@@ -1102,14 +1102,14 @@ EOF;
 '~': 'null'
 EOF;
 
-        $expected = array(
+        $expected = [
             '1.2' => 'bar',
             '1.3' => 'baz',
             'true' => 'foo',
             'false' => 'bar',
             'null' => 'null',
             '~' => 'null',
-        );
+        ];
 
         $this->assertEquals($expected, $this->parser->parse($yaml));
     }
@@ -1134,7 +1134,7 @@ foo:
     bar: foobar # Note: a comment after a colon
 EOT;
 
-        $this->assertSame(array('foo' => array('bar' => 'foobar')), $this->parser->parse($yaml));
+        $this->assertSame(['foo' => ['bar' => 'foobar']], $this->parser->parse($yaml));
     }
 
     /**
@@ -1147,7 +1147,7 @@ EOT;
 
     public function getCommentLikeStringInScalarBlockData()
     {
-        $tests = array();
+        $tests = [];
 
         $yaml = <<<'EOT'
 pages:
@@ -1164,9 +1164,9 @@ pages:
 
             footer # comment3
 EOT;
-        $expected = array(
-            'pages' => array(
-                array(
+        $expected = [
+            'pages' => [
+                [
                     'title' => 'some title',
                     'content' => <<<'EOT'
 # comment 1
@@ -1180,10 +1180,10 @@ header
 footer # comment3
 EOT
                     ,
-                ),
-            ),
-        );
-        $tests[] = array($yaml, $expected);
+                ],
+            ],
+        ];
+        $tests[] = [$yaml, $expected];
 
         $yaml = <<<'EOT'
 test: |
@@ -1200,7 +1200,7 @@ collection:
         # bar
         baz
 EOT;
-        $expected = array(
+        $expected = [
             'test' => <<<'EOT'
 foo
 # bar
@@ -1208,8 +1208,8 @@ baz
 
 EOT
             ,
-            'collection' => array(
-                array(
+            'collection' => [
+                [
                     'one' => <<<'EOT'
 foo
 # bar
@@ -1217,18 +1217,18 @@ baz
 
 EOT
                     ,
-                ),
-                array(
+                ],
+                [
                     'two' => <<<'EOT'
 foo
 # bar
 baz
 EOT
                     ,
-                ),
-            ),
-        );
-        $tests[] = array($yaml, $expected);
+                ],
+            ],
+        ];
+        $tests[] = [$yaml, $expected];
 
         $yaml = <<<'EOT'
 foo:
@@ -1240,17 +1240,17 @@ foo:
 # comment
     foobar: ~
 EOT;
-        $expected = array(
-            'foo' => array(
-                'bar' => array(
+        $expected = [
+            'foo' => [
+                'bar' => [
                     'scalar-block' => "line1 line2>\n",
-                ),
-                'baz' => array(
+                ],
+                'baz' => [
                     'foobar' => null,
-                ),
-            ),
-        );
-        $tests[] = array($yaml, $expected);
+                ],
+            ],
+        ];
+        $tests[] = [$yaml, $expected];
 
         $yaml = <<<'EOT'
 a:
@@ -1260,13 +1260,13 @@ a:
 #        second row
     d: hello
 EOT;
-        $expected = array(
-            'a' => array(
+        $expected = [
+            'a' => [
                 'b' => 'hello',
                 'd' => 'hello',
-            ),
-        );
-        $tests[] = array($yaml, $expected);
+            ],
+        ];
+        $tests[] = [$yaml, $expected];
 
         return $tests;
     }
@@ -1284,13 +1284,13 @@ test: >
 EOT;
 
         $this->assertSame(
-            array(
+            [
                 'test' => <<<'EOT'
 <h2>A heading</h2>
 <ul> <li>a list</li> <li>may be a good example</li> </ul>
 EOT
                 ,
-            ),
+            ],
             $this->parser->parse($yaml)
         );
     }
@@ -1308,7 +1308,7 @@ test: >
 EOT;
 
         $this->assertSame(
-            array(
+            [
                 'test' => <<<'EOT'
 <h2>A heading</h2>
 <ul>
@@ -1317,7 +1317,7 @@ EOT;
 </ul>
 EOT
                 ,
-            ),
+            ],
             $this->parser->parse($yaml)
         );
     }
@@ -1327,28 +1327,28 @@ EOT
      */
     public function testParseBinaryData($data)
     {
-        $this->assertSame(array('data' => 'Hello world'), $this->parser->parse($data));
+        $this->assertSame(['data' => 'Hello world'], $this->parser->parse($data));
     }
 
     public function getBinaryData()
     {
-        return array(
-            'enclosed with double quotes' => array('data: !!binary "SGVsbG8gd29ybGQ="'),
-            'enclosed with single quotes' => array("data: !!binary 'SGVsbG8gd29ybGQ='"),
-            'containing spaces' => array('data: !!binary  "SGVs bG8gd 29ybGQ="'),
-            'in block scalar' => array(
+        return [
+            'enclosed with double quotes' => ['data: !!binary "SGVsbG8gd29ybGQ="'],
+            'enclosed with single quotes' => ["data: !!binary 'SGVsbG8gd29ybGQ='"],
+            'containing spaces' => ['data: !!binary  "SGVs bG8gd 29ybGQ="'],
+            'in block scalar' => [
                 <<<'EOT'
 data: !!binary |
     SGVsbG8gd29ybGQ=
 EOT
-    ),
-            'containing spaces in block scalar' => array(
+    ],
+            'containing spaces in block scalar' => [
                 <<<'EOT'
 data: !!binary |
     SGVs bG8gd 29ybGQ=
 EOT
-    ),
-        );
+    ],
+        ];
     }
 
     /**
@@ -1368,44 +1368,44 @@ EOT
 
     public function getInvalidBinaryData()
     {
-        return array(
-            'length not a multiple of four' => array('data: !!binary "SGVsbG8d29ybGQ="', '/The normalized base64 encoded data \(data without whitespace characters\) length must be a multiple of four \(\d+ bytes given\)/'),
-            'invalid characters' => array('!!binary "SGVsbG8#d29ybGQ="', '/The base64 encoded data \(.*\) contains invalid characters/'),
-            'too many equals characters' => array('data: !!binary "SGVsbG8gd29yb==="', '/The base64 encoded data \(.*\) contains invalid characters/'),
-            'misplaced equals character' => array('data: !!binary "SGVsbG8gd29ybG=Q"', '/The base64 encoded data \(.*\) contains invalid characters/'),
-            'length not a multiple of four in block scalar' => array(
+        return [
+            'length not a multiple of four' => ['data: !!binary "SGVsbG8d29ybGQ="', '/The normalized base64 encoded data \(data without whitespace characters\) length must be a multiple of four \(\d+ bytes given\)/'],
+            'invalid characters' => ['!!binary "SGVsbG8#d29ybGQ="', '/The base64 encoded data \(.*\) contains invalid characters/'],
+            'too many equals characters' => ['data: !!binary "SGVsbG8gd29yb==="', '/The base64 encoded data \(.*\) contains invalid characters/'],
+            'misplaced equals character' => ['data: !!binary "SGVsbG8gd29ybG=Q"', '/The base64 encoded data \(.*\) contains invalid characters/'],
+            'length not a multiple of four in block scalar' => [
                 <<<'EOT'
 data: !!binary |
     SGVsbG8d29ybGQ=
 EOT
                 ,
                 '/The normalized base64 encoded data \(data without whitespace characters\) length must be a multiple of four \(\d+ bytes given\)/',
-            ),
-            'invalid characters in block scalar' => array(
+            ],
+            'invalid characters in block scalar' => [
                 <<<'EOT'
 data: !!binary |
     SGVsbG8#d29ybGQ=
 EOT
                 ,
                 '/The base64 encoded data \(.*\) contains invalid characters/',
-            ),
-            'too many equals characters in block scalar' => array(
+            ],
+            'too many equals characters in block scalar' => [
                 <<<'EOT'
 data: !!binary |
     SGVsbG8gd29yb===
 EOT
                 ,
                 '/The base64 encoded data \(.*\) contains invalid characters/',
-            ),
-            'misplaced equals character in block scalar' => array(
+            ],
+            'misplaced equals character in block scalar' => [
                 <<<'EOT'
 data: !!binary |
     SGVsbG8gd29ybG=Q
 EOT
                 ,
                 '/The base64 encoded data \(.*\) contains invalid characters/',
-            ),
-        );
+            ],
+        ];
     }
 
     public function testParseDateAsMappingValue()
@@ -1418,7 +1418,7 @@ EOT;
         $expectedDate->setDate(2002, 12, 14);
         $expectedDate->setTime(0, 0, 0);
 
-        $this->assertEquals(array('date' => $expectedDate), $this->parser->parse($yaml, Yaml::PARSE_DATETIME));
+        $this->assertEquals(['date' => $expectedDate], $this->parser->parse($yaml, Yaml::PARSE_DATETIME));
     }
 
     /**
@@ -1440,8 +1440,8 @@ EOT;
 
     public function parserThrowsExceptionWithCorrectLineNumberProvider()
     {
-        return array(
-            array(
+        return [
+            [
                 4,
                 <<<'YAML'
 foo:
@@ -1449,8 +1449,8 @@ foo:
         # bar
         bar: "123",
 YAML
-            ),
-            array(
+            ],
+            [
                 5,
                 <<<'YAML'
 foo:
@@ -1459,8 +1459,8 @@ foo:
         # bar
         bar: "123",
 YAML
-            ),
-            array(
+            ],
+            [
                 8,
                 <<<'YAML'
 foo:
@@ -1472,8 +1472,8 @@ bar:
         # bar
         bar: "123",
 YAML
-            ),
-            array(
+            ],
+            [
                 10,
                 <<<'YAML'
 foo:
@@ -1487,8 +1487,8 @@ bar:
         # bar
         bar: "123",
 YAML
-            ),
-        );
+            ],
+        ];
     }
 
     public function testParseMultiLineQuotedString()
@@ -1501,7 +1501,7 @@ foo"
 bar: baz
 EOT;
 
-        $this->assertSame(array('foo' => 'bar baz foobar foo', 'bar' => 'baz'), $this->parser->parse($yaml));
+        $this->assertSame(['foo' => 'bar baz foobar foo', 'bar' => 'baz'], $this->parser->parse($yaml));
     }
 
     public function testMultiLineQuotedStringWithTrailingBackslash()
@@ -1512,7 +1512,7 @@ foobar:
     bar"
 YAML;
 
-        $this->assertSame(array('foobar' => 'foobar'), $this->parser->parse($yaml));
+        $this->assertSame(['foobar' => 'foobar'], $this->parser->parse($yaml));
     }
 
     public function testCommentCharactersInMultiLineQuotedStrings()
@@ -1523,12 +1523,12 @@ foo:
       #bar'
     bar: baz
 YAML;
-        $expected = array(
-            'foo' => array(
+        $expected = [
+            'foo' => [
                 'foobar' => 'foo #bar',
                 'bar' => 'baz',
-            ),
-        );
+            ],
+        ];
 
         $this->assertSame($expected, $this->parser->parse($yaml));
     }
@@ -1540,9 +1540,9 @@ foobar: 'foo
 
     bar'
 YAML;
-        $expected = array(
+        $expected = [
             'foobar' => "foo\nbar",
-        );
+        ];
 
         $this->assertSame($expected, $this->parser->parse($yaml));
     }
@@ -1557,7 +1557,7 @@ foo: bar
 bar: baz
 EOT;
 
-        $this->assertSame(array('foo' => 'bar baz foobar foo', 'bar' => 'baz'), $this->parser->parse($yaml));
+        $this->assertSame(['foo' => 'bar baz foobar foo', 'bar' => 'baz'], $this->parser->parse($yaml));
     }
 
     public function testParseMultiLineString()
@@ -1575,7 +1575,7 @@ EOT;
 
     public function multiLineDataProvider()
     {
-        $tests = array();
+        $tests = [];
 
         $yaml = <<<'EOF'
 foo:
@@ -1585,15 +1585,15 @@ foo:
     two
     three
 EOF;
-        $expected = array(
-            'foo' => array(
-                array(
+        $expected = [
+            'foo' => [
+                [
                     'bar' => "one\ntwo three",
-                ),
-            ),
-        );
+                ],
+            ],
+        ];
 
-        $tests[] = array($yaml, $expected, false);
+        $tests[] = [$yaml, $expected, false];
 
         $yaml = <<<'EOF'
 bar
@@ -1601,7 +1601,7 @@ bar
 EOF;
         $expected = 'bar "foo"';
 
-        $tests[] = array($yaml, $expected, false);
+        $tests[] = [$yaml, $expected, false];
 
         $yaml = <<<'EOF'
 bar
@@ -1609,7 +1609,7 @@ bar
 EOF;
         $expected = 'bar "foo';
 
-        $tests[] = array($yaml, $expected, false);
+        $tests[] = [$yaml, $expected, false];
 
         $yaml = <<<'EOF'
 bar
@@ -1618,7 +1618,7 @@ bar
 EOF;
         $expected = "bar\n'foo'";
 
-        $tests[] = array($yaml, $expected, false);
+        $tests[] = [$yaml, $expected, false];
 
         $yaml = <<<'EOF'
 bar
@@ -1627,14 +1627,14 @@ foo'
 EOF;
         $expected = "bar\nfoo'";
 
-        $tests[] = array($yaml, $expected, false);
+        $tests[] = [$yaml, $expected, false];
 
         return $tests;
     }
 
     public function testTaggedInlineMapping()
     {
-        $this->assertEquals(new TaggedValue('foo', array('foo' => 'bar')), $this->parser->parse('!foo {foo: bar}', Yaml::PARSE_CUSTOM_TAGS));
+        $this->assertEquals(new TaggedValue('foo', ['foo' => 'bar']), $this->parser->parse('!foo {foo: bar}', Yaml::PARSE_CUSTOM_TAGS));
     }
 
     /**
@@ -1647,48 +1647,48 @@ EOF;
 
     public function taggedValuesProvider()
     {
-        return array(
-            'scalars' => array(
-                array(
+        return [
+            'scalars' => [
+                [
                     'foo' => new TaggedValue('inline', 'bar'),
                     'quz' => new TaggedValue('long', 'this is a long text'),
-                ),
+                ],
                 <<<YAML
 foo: !inline bar
 quz: !long >
   this is a long
   text
 YAML
-            ),
-            'sequences' => array(
-                array(new TaggedValue('foo', array('yaml')), new TaggedValue('quz', array('bar'))),
+            ],
+            'sequences' => [
+                [new TaggedValue('foo', ['yaml']), new TaggedValue('quz', ['bar'])],
                 <<<YAML
 - !foo
     - yaml
 - !quz [bar]
 YAML
-            ),
-            'mappings' => array(
-                new TaggedValue('foo', array('foo' => new TaggedValue('quz', array('bar')), 'quz' => new TaggedValue('foo', array('quz' => 'bar')))),
+            ],
+            'mappings' => [
+                new TaggedValue('foo', ['foo' => new TaggedValue('quz', ['bar']), 'quz' => new TaggedValue('foo', ['quz' => 'bar'])]),
                 <<<YAML
 !foo
 foo: !quz [bar]
 quz: !foo
    quz: bar
 YAML
-            ),
-            'inline' => array(
-                array(new TaggedValue('foo', array('foo', 'bar')), new TaggedValue('quz', array('foo' => 'bar', 'quz' => new TaggedValue('bar', array('one' => 'bar'))))),
+            ],
+            'inline' => [
+                [new TaggedValue('foo', ['foo', 'bar']), new TaggedValue('quz', ['foo' => 'bar', 'quz' => new TaggedValue('bar', ['one' => 'bar'])])],
                 <<<YAML
 - !foo [foo, bar]
 - !quz {foo: bar, quz: !bar {one: bar}}
 YAML
-            ),
-            'spaces-around-tag-value-in-sequence' => array(
-                array(new TaggedValue('foo', 'bar')),
+            ],
+            'spaces-around-tag-value-in-sequence' => [
+                [new TaggedValue('foo', 'bar')],
                 '[ !foo bar ]',
-            ),
-        );
+            ],
+        ];
     }
 
     public function testNonSpecificTagSupport()
@@ -1797,7 +1797,7 @@ INI;
     {
         $parser = new Parser();
 
-        $tests = array();
+        $tests = [];
         $files = $parser->parseFile(__DIR__.'/Fixtures/'.$testsFile);
         foreach ($files as $file) {
             $yamls = file_get_contents(__DIR__.'/Fixtures/'.$file.'.yml');
@@ -1814,7 +1814,7 @@ INI;
                 } else {
                     eval('$expected = '.trim($test['php']).';');
 
-                    $tests[] = array(var_export($expected, true), $test['yaml'], $test['test']);
+                    $tests[] = [var_export($expected, true), $test['yaml'], $test['test']];
                 }
             }
         }
@@ -1825,7 +1825,7 @@ INI;
     public function testCanParseVeryLongValue()
     {
         $longStringWithSpaces = str_repeat('xxxxxxxxxxxxxxxxxxxxxxxxxxxxxx ', 20000);
-        $trickyVal = array('x' => $longStringWithSpaces);
+        $trickyVal = ['x' => $longStringWithSpaces];
 
         $yamlString = Yaml::dump($trickyVal);
         $arrayFromYaml = $this->parser->parse($yamlString);
@@ -1863,16 +1863,16 @@ transitions:
             - !php/const 'Symfony\Component\Yaml\Tests\B::BAR'
         to: !php/const 'Symfony\Component\Yaml\Tests\B::BAZ'
 YAML;
-        $expected = array(
-            'transitions' => array(
-                'foo' => array(
-                    'from' => array(
+        $expected = [
+            'transitions' => [
+                'foo' => [
+                    'from' => [
                         'bar',
-                    ),
+                    ],
                     'to' => 'baz',
-                ),
-            ),
-        );
+                ],
+            ],
+        ];
 
         $this->assertSame($expected, $this->parser->parse($yaml, Yaml::PARSE_CONSTANT));
     }
@@ -1893,23 +1893,23 @@ foobar:
     bar: ~
     <<: [*FOO, *BAR]
 YAML;
-        $expected = (object) array(
-            'foo' => (object) array(
+        $expected = (object) [
+            'foo' => (object) [
                 'bar' => 1,
-            ),
-            'bar' => (object) array(
+            ],
+            'bar' => (object) [
                 'baz' => 2,
                 'bar' => 1,
-            ),
-            'baz' => (object) array(
+            ],
+            'baz' => (object) [
                 'baz_foo' => 3,
                 'baz_bar' => 4,
-            ),
-            'foobar' => (object) array(
+            ],
+            'foobar' => (object) [
                 'bar' => null,
                 'baz' => 2,
-            ),
-        );
+            ],
+        ];
 
         $this->assertEquals($expected, $this->parser->parse($yaml, Yaml::PARSE_OBJECT_FOR_MAP));
     }
@@ -1967,18 +1967,18 @@ mergekeyderef:
     d: quux
     <<: *quux
 YAML;
-        $expected = array(
-            'mergekeyrefdef' => array(
+        $expected = [
+            'mergekeyrefdef' => [
                 'a' => 'foo',
                 'b' => 'bar',
                 'c' => 'baz',
-            ),
-            'mergekeyderef' => array(
+            ],
+            'mergekeyderef' => [
                 'd' => 'quux',
                 'b' => 'bar',
                 'c' => 'baz',
-            ),
-        );
+            ],
+        ];
 
         $this->assertSame($expected, $this->parser->parse($yaml));
     }
@@ -1995,18 +1995,18 @@ mergekeyderef:
     d: quux
     <<: *quux
 YAML;
-        $expected = (object) array(
-            'mergekeyrefdef' => (object) array(
+        $expected = (object) [
+            'mergekeyrefdef' => (object) [
                 'a' => 'foo',
                 'b' => 'bar',
                 'c' => 'baz',
-            ),
-            'mergekeyderef' => (object) array(
+            ],
+            'mergekeyderef' => (object) [
                 'd' => 'quux',
                 'b' => 'bar',
                 'c' => 'baz',
-            ),
-        );
+            ],
+        ];
 
         $this->assertEquals($expected, $this->parser->parse($yaml, Yaml::PARSE_OBJECT_FOR_MAP));
     }
@@ -2024,6 +2024,48 @@ EOE;
     }
 
     /**
+     * @dataProvider circularReferenceProvider
+     * @expectedException \Symfony\Component\Yaml\Exception\ParseException
+     * @expectedExceptionMessage Circular reference [foo, bar, foo] detected
+     */
+    public function testDetectCircularReferences($yaml)
+    {
+        $this->parser->parse($yaml, Yaml::PARSE_CUSTOM_TAGS);
+    }
+
+    public function circularReferenceProvider()
+    {
+        $tests = [];
+
+        $yaml = <<<YAML
+foo:
+    - &foo
+      - &bar
+        bar: foobar
+        baz: *foo
+YAML;
+        $tests['sequence'] = [$yaml];
+
+        $yaml = <<<YAML
+foo: &foo
+    bar: &bar
+        foobar: baz
+        baz: *foo
+YAML;
+        $tests['mapping'] = [$yaml];
+
+        $yaml = <<<YAML
+foo: &foo
+    bar: &bar
+        foobar: baz
+        <<: *foo
+YAML;
+        $tests['mapping with merge key'] = [$yaml];
+
+        return $tests;
+    }
+
+    /**
      * @dataProvider indentedMappingData
      */
     public function testParseIndentedMappings($yaml, $expected)
@@ -2033,7 +2075,7 @@ EOE;
 
     public function indentedMappingData()
     {
-        $tests = array();
+        $tests = [];
 
         $yaml = <<<YAML
 foo:
@@ -2041,15 +2083,15 @@ foo:
     # A comment
     baz: "foobaz"
 YAML;
-        $expected = array(
-            'foo' => array(
-                array(
+        $expected = [
+            'foo' => [
+                [
                     'bar' => 'foobar',
                     'baz' => 'foobaz',
-                ),
-            ),
-        );
-        $tests['comment line is first line in indented block'] = array($yaml, $expected);
+                ],
+            ],
+        ];
+        $tests['comment line is first line in indented block'] = [$yaml, $expected];
 
         $yaml = <<<YAML
 foo:
@@ -2057,42 +2099,42 @@ foo:
         # comment
         baz: [1, 2, 3]
 YAML;
-        $expected = array(
-            'foo' => array(
-                array(
-                    'bar' => array(
-                        'baz' => array(1, 2, 3),
-                    ),
-                ),
-            ),
-        );
-        $tests['mapping value on new line starting with a comment line'] = array($yaml, $expected);
+        $expected = [
+            'foo' => [
+                [
+                    'bar' => [
+                        'baz' => [1, 2, 3],
+                    ],
+                ],
+            ],
+        ];
+        $tests['mapping value on new line starting with a comment line'] = [$yaml, $expected];
 
         $yaml = <<<YAML
 foo:
   -
     bar: foobar
 YAML;
-        $expected = array(
-            'foo' => array(
-                array(
+        $expected = [
+            'foo' => [
+                [
                     'bar' => 'foobar',
-                ),
-            ),
-        );
-        $tests['mapping in sequence starting on a new line'] = array($yaml, $expected);
+                ],
+            ],
+        ];
+        $tests['mapping in sequence starting on a new line'] = [$yaml, $expected];
 
         $yaml = <<<YAML
 foo:
 
     bar: baz
 YAML;
-        $expected = array(
-            'foo' => array(
+        $expected = [
+            'foo' => [
                 'bar' => 'baz',
-            ),
-        );
-        $tests['blank line at the beginning of an indented mapping value'] = array($yaml, $expected);
+            ],
+        ];
+        $tests['blank line at the beginning of an indented mapping value'] = [$yaml, $expected];
 
         return $tests;
     }
