@@ -7,12 +7,70 @@ $baseDir = dirname($vendorDir);
 
 return array(
     'phpDocumentor\\Reflection\\' => array($vendorDir . '/phpdocumentor/reflection-common/src', $vendorDir . '/phpdocumentor/type-resolver/src', $vendorDir . '/phpdocumentor/reflection-docblock/src'),
-    'Zend\\' => array($vendorDir . '/zendframework/zendframework/library/Zend'),
+    'Zend\\XmlRpc\\' => array($vendorDir . '/zendframework/zend-xmlrpc/src'),
+    'Zend\\View\\' => array($vendorDir . '/zendframework/zend-view/src'),
+    'Zend\\Version\\' => array($vendorDir . '/zendframework/zend-version/src'),
+    'Zend\\Validator\\' => array($vendorDir . '/zendframework/zend-validator/src'),
+    'Zend\\Uri\\' => array($vendorDir . '/zendframework/zend-uri/src'),
+    'Zend\\Text\\' => array($vendorDir . '/zendframework/zend-text/src'),
+    'Zend\\Test\\' => array($vendorDir . '/zendframework/zend-test/src'),
+    'Zend\\Tag\\' => array($vendorDir . '/zendframework/zend-tag/src'),
+    'Zend\\Stdlib\\' => array($vendorDir . '/zendframework/zend-stdlib/src'),
+    'Zend\\Soap\\' => array($vendorDir . '/zendframework/zend-soap/src'),
+    'Zend\\Session\\' => array($vendorDir . '/zendframework/zend-session/src'),
+    'Zend\\ServiceManager\\' => array($vendorDir . '/zendframework/zend-servicemanager/src'),
+    'Zend\\Server\\' => array($vendorDir . '/zendframework/zend-server/src'),
+    'Zend\\Serializer\\' => array($vendorDir . '/zendframework/zend-serializer/src'),
+    'Zend\\Psr7Bridge\\' => array($vendorDir . '/zendframework/zend-psr7bridge/src'),
+    'Zend\\ProgressBar\\' => array($vendorDir . '/zendframework/zend-progressbar/src'),
+    'Zend\\Permissions\\Rbac\\' => array($vendorDir . '/zendframework/zend-permissions-rbac/src'),
+    'Zend\\Permissions\\Acl\\' => array($vendorDir . '/zendframework/zend-permissions-acl/src'),
+    'Zend\\Paginator\\' => array($vendorDir . '/zendframework/zend-paginator/src'),
+    'Zend\\Navigation\\' => array($vendorDir . '/zendframework/zend-navigation/src'),
+    'Zend\\Mvc\\' => array($vendorDir . '/zendframework/zend-mvc/src'),
+    'Zend\\ModuleManager\\' => array($vendorDir . '/zendframework/zend-modulemanager/src'),
+    'Zend\\Mime\\' => array($vendorDir . '/zendframework/zend-mime/src'),
+    'Zend\\Memory\\' => array($vendorDir . '/zendframework/zend-memory/src'),
+    'Zend\\Math\\' => array($vendorDir . '/zendframework/zend-math/src'),
+    'Zend\\Mail\\' => array($vendorDir . '/zendframework/zend-mail/src'),
+    'Zend\\Log\\' => array($vendorDir . '/zendframework/zend-log/src'),
+    'Zend\\Loader\\' => array($vendorDir . '/zendframework/zend-loader/src'),
+    'Zend\\Json\\' => array($vendorDir . '/zendframework/zend-json/src'),
+    'Zend\\InputFilter\\' => array($vendorDir . '/zendframework/zend-inputfilter/src'),
+    'Zend\\I18n\\' => array($vendorDir . '/zendframework/zend-i18n/src'),
+    'Zend\\Hydrator\\' => array($vendorDir . '/zendframework/zend-hydrator/src'),
+    'Zend\\Http\\' => array($vendorDir . '/zendframework/zend-http/src'),
+    'Zend\\Form\\' => array($vendorDir . '/zendframework/zend-form/src'),
+    'Zend\\Filter\\' => array($vendorDir . '/zendframework/zend-filter/src'),
+    'Zend\\File\\' => array($vendorDir . '/zendframework/zend-file/src'),
+    'Zend\\Feed\\' => array($vendorDir . '/zendframework/zend-feed/src'),
+    'Zend\\EventManager\\' => array($vendorDir . '/zendframework/zend-eventmanager/src'),
+    'Zend\\Escaper\\' => array($vendorDir . '/zendframework/zend-escaper/src'),
+    'Zend\\Dom\\' => array($vendorDir . '/zendframework/zend-dom/src'),
+    'Zend\\Diactoros\\' => array($vendorDir . '/zendframework/zend-diactoros/src'),
+    'Zend\\Di\\' => array($vendorDir . '/zendframework/zend-di/src'),
+    'Zend\\Debug\\' => array($vendorDir . '/zendframework/zend-debug/src'),
+    'Zend\\Db\\' => array($vendorDir . '/zendframework/zend-db/src'),
+    'Zend\\Crypt\\' => array($vendorDir . '/zendframework/zend-crypt/src'),
+    'Zend\\Console\\' => array($vendorDir . '/zendframework/zend-console/src'),
+    'Zend\\Config\\' => array($vendorDir . '/zendframework/zend-config/src'),
+    'Zend\\Code\\' => array($vendorDir . '/zendframework/zend-code/src'),
+    'Zend\\Captcha\\' => array($vendorDir . '/zendframework/zend-captcha/src'),
+    'Zend\\Cache\\' => array($vendorDir . '/zendframework/zend-cache/src'),
+    'Zend\\Barcode\\' => array($vendorDir . '/zendframework/zend-barcode/src'),
+    'Zend\\Authentication\\' => array($vendorDir . '/zendframework/zend-authentication/src'),
     'ZendXml\\' => array($vendorDir . '/zendframework/zendxml/src'),
     'Webmozart\\Assert\\' => array($vendorDir . '/webmozart/assert/src'),
+    'TrueBV\\' => array($vendorDir . '/true/punycode/src'),
+    'Symfony\\Polyfill\\Mbstring\\' => array($vendorDir . '/symfony/polyfill-mbstring'),
     'Symfony\\Polyfill\\Ctype\\' => array($vendorDir . '/symfony/polyfill-ctype'),
     'Symfony\\Component\\Yaml\\' => array($vendorDir . '/symfony/yaml'),
+    'Psr\\SimpleCache\\' => array($vendorDir . '/psr/simple-cache/src'),
+    'Psr\\Log\\' => array($vendorDir . '/psr/log/Psr/Log'),
     'Psr\\Http\\Message\\' => array($vendorDir . '/psr/http-message/src'),
+    'Psr\\Container\\' => array($vendorDir . '/psr/container/src'),
+    'Psr\\Cache\\' => array($vendorDir . '/psr/cache/src'),
+    'Interop\\Container\\' => array($vendorDir . '/container-interop/container-interop/src/Interop/Container'),
     'GuzzleHttp\\Psr7\\' => array($vendorDir . '/guzzlehttp/psr7/src'),
     'GuzzleHttp\\Promise\\' => array($vendorDir . '/guzzlehttp/promises/src'),
     'GuzzleHttp\\' => array($vendorDir . '/guzzlehttp/guzzle/src'),
